Struct qd::Double

pub struct Double(_, _);

A 128-bit floating-point number implemented as the unevaluated sum of two 64-bit floating-point numbers.

There are several ways to create a new Double:

calling the new or norm functions

calling Double::from and passing a type that has a From implementation

calling parse on a string

calling from_add, from_sub, from_mul, or from_div

using the dd macro

See the module-level documentation (index.html) for more information.

Methods

impl Double

pub const RADIX: u32

The radix or base of the internal representation of Double. This is the same as the representation in the underlying f64.

pub const MANTISSA_DIGITS: u32

Number of significant digits in base 2.

pub const DIGITS: u32

Approximate number of significant digits in base 10.

pub const EPSILON: Double

Machine epsilon value for Double.

This is the difference between 1.0 and the next largest representable number.

pub const MIN: Double

Smallest finite Double value.

pub const MIN_POSITIVE: Double

Smallest positive normal Double value.

pub const MAX: Double

Largest finite Double value.

pub const MIN_EXP: i32

One greater than the minimum possible normal power of 2 exponent.

pub const MAX_EXP: i32

Maximum possible power of 2 exponent.

pub const MIN_10_EXP: i32

Minimum possible normal power of 10 exponent.

pub const MAX_10_EXP: i32

Maximum possible power of 10 exponent.

pub const NAN: Double

Not a Number (NaN).

pub const INFINITY: Double

Infinity (∞).

pub const NEG_INFINITY: Double

Negative infinity (-∞).

pub const ZERO: Double

Zero (0)

pub const NEG_ZERO: Double

Negative zero (-0)

pub const ONE: Double

One (1)

pub const PI: Double

Archimedes' constant (π)

pub const MUL_2_PI: Double

2π

pub const FRAC_PI_2: Double

π/2

pub const FRAC_PI_3: Double

π/3

pub const FRAC_PI_4: Double

π/4

pub const FRAC_PI_6: Double

π/6

pub const FRAC_PI_8: Double

π/8

pub const FRAC_PI_16: Double

π/16

pub const FRAC_3_PI_4: Double

3π/4

pub const FRAC_1_PI: Double

1/π

pub const FRAC_2_PI: Double

2/π

pub const FRAC_2_SQRT_PI: Double

2/√π

pub const SQRT_2: Double

√2

pub const FRAC_1_SQRT_2: Double

1/√2

pub const E: Double

Euler's number (e)

pub const LOG2_10: Double

log₂ 10

pub const LOG2_E: Double

log₂ e

pub const LOG10_2: Double

log 2

pub const LOG10_E: Double

log e

pub const LN_2: Double

ln 2

pub const LN_10: Double

ln 10

impl Double

pub fn from_add(a: f64, b: f64) -> Double

Creates a new double-double representing the sum of two floats.

Examples

let x = Double::from_add(1.0, 2.0);
assert!(x == dd!(3.0));

impl Double

pub fn from_sub(a: f64, b: f64) -> Double

Creates a new double-double representing the difference of two floats.

Examples

let x = Double::from_sub(1.0, 2.0);
assert!(x == dd!(-1.0));

impl Double

pub fn from_mul(a: f64, b: f64) -> Double

Creates a new double-double representing the product of two floats.

Examples

let x = Double::from_mul(1.0, 2.0);
assert!(x == dd!(2.0));

impl Double

pub fn from_div(a: f64, b: f64) -> Double

Creates a new double-double representing the quotient of two floats.

Examples

let x = Double::from_div(1.0, 2.0);
assert!(x == dd!(0.5));

pub fn recip(self) -> Double

Calculates the reciprocal of the number.

Examples

let r = Double::from(2.0).recip();
assert!(r == dd!(0.5));

impl Double

pub fn as_float(self) -> f64

Converts the number into an f64.

pub fn as_int(self) -> i64

Converts the number into an i64.

While it is possible for a Double to be created from a u64, whether or not the original is signed is not recorded (since Double itself is signed). The return value of this function can be cast to u64 if necessary.

impl Double

pub fn sqr(self) -> Double

Calculates the square of the number.

This method takes advantage of optimizations in multiplication that are available when the two numbers being multiplied are the same, so it is more efficient than bare multiplication.

Examples

let x = dd!(3);
assert!(x.sqr() == x * x); // The left side is faster though

pub fn powi(self, n: i32) -> Double

Calculates the number raised to an integral power.

Examples

let x = dd!(3);
assert!(x.powi(3) == dd!(27));

pub fn powf(self, n: Double) -> Double

Calculates the number raised to a double-double power.

This function only works for positive values of the number as it uses a simplified logarithm-based algorithm. This is likely to change in the future when a more complex algorithm is implemented.

Examples

let x = dd!(3).powf(dd!(3.3));
let expected = dd!("37.540507598529552193101865954634");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn ldexp(self, n: i32) -> Double

Calculates self × 2n and returns it as a new Double.

Though this is not an everyday operation, it is often used in more advanced mathematical calculations (including several within this library). Therefore an implementation that is much more efficient than calculating it through multiplication and powi is offered despite it not being part of the f64 API.

Examples

let x = dd!(3);
assert!(x.ldexp(3) == dd!(24)); // 3 * 2^3

impl Double

pub fn sqrt(self) -> Double

Calculates the square root of the number.

Examples

let x = dd!(2).sqrt();
let diff = (x - Double::SQRT_2).abs();
assert!(diff < dd!(1e-30));

pub fn cbrt(self) -> Double

Calculates the cube root of the number.

Examples

let x = dd!(2).cbrt();
let expected = dd!("1.2599210498948731647672106072782");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn nroot(self, n: i32) -> Double

Calculates the nth root of the number.

Examples

let x = dd!(2).nroot(4);
let expected = dd!("1.1892071150027210667174999705605");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

impl Double

pub fn exp(self) -> Double

Computes the exponential function, ex, for the number.

Examples

let x = dd!(2).exp();
let expected = dd!("7.3890560989306502272304274605750");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

impl Double

pub fn ln(self) -> Double

Calculates the natural logarithm, loge, of the number.

Examples

let x = Double::from(7).ln();
let expected = dd!("1.9459101490553133051053527434432");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn log10(self) -> Double

Calculates log10 of the number.

Examples

let x = Double::E.log10();
let expected = dd!("0.434294481903251827651128918916605");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn log2(self) -> Double

Calculates log2 of the number.

Examples

let x = dd!(10).log2();
let expected = dd!("3.32192809488736234787031942948939");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn log(self, b: f64) -> Double

Calculates the base b logarithm (logb) of the number.

If the goal is to calculate the base e, base 2, or base 10 logarithms of self, the specialized functions for those purposes(ln, log2, and log10 respectively) will be more efficient.

Examples

let x = dd!(10).log(7.0);
let expected = dd!("1.18329466245493832681792856164686");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

impl Double

pub fn sin(self) -> Double

Computes the sine of the number.

Examples

let x = (Double::PI / dd!(2)).sin();
let expected = dd!(1);

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn cos(self) -> Double

Computes the cosine of the number.

Examples

let x = (Double::PI / dd!(2)).cos();
let expected = dd!(0);

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn sin_cos(self) -> (Double, Double)

Simultaneously computes the sine and the cosine of the number. This is more efficient if you need both numbers. Returns (sin(x), cos(x)).

Examples

let x = Double::PI / dd!(4);
let (sin_x, cos_x) = x.sin_cos();

let diff_sin = (sin_x - x.sin()).abs();
let diff_cos = (cos_x - x.cos()).abs();

assert!(diff_sin < dd!(1e-30));
assert!(diff_cos < dd!(1e-30));

pub fn tan(self) -> Double

Computes the tangent of the number.

Examples

let x = (Double::PI / dd!(4)).tan();
let expected = dd!(1);

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn atan2(self, other: Double) -> Double

Computes the 2-argument arctangent of the number (y) and other (x) in radians.

The second argument allows the avoidance of ambiguities in the single-argument atan function, notably allowing the determination of quadrant.

Examples

let pi = Double::PI;

// -π/4 radians (45 degrees clockwise)
let x1 = dd!(3);
let y1 = dd!(-3);
let expected1 = -pi / dd!(4);

// 3π/4 radians (135 degrees counter-clockwise)
let x2 = dd!(-3);
let y2 = dd!(3);
let expected2 = Double::from_div(3.0, 4.0) * pi;

let diff1 = (y1.atan2(x1) - expected1).abs();
let diff2 = (y2.atan2(x2) - expected2).abs();

assert!(diff1 < dd!(1e-30));
assert!(diff2 < dd!(1e-30));

pub fn atan(self) -> Double

Computes the arctangent of the number. The return value is in the range [-π/2, π/2].

Examples

let x = dd!(1).atan();
let expected = Double::PI / dd!(4);  // π/4

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn asin(self) -> Double

Computes the arcsine of the number. The return value is in the range [-π/2, π/2] for any number in the range [-1, 1]. Otherwise the return value is NaN.

Examples

let x = dd!(1).asin();
let expected = Double::PI / dd!(2);  // π/2

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn acos(self) -> Double

Computes the arccosine of the number. The return value is in the range [0, π] for any number in the range [-1, 1]. Otherwise the return value is NaN.

Examples

let x = dd!(1).acos();
let expected = dd!(0);

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

impl Double

pub fn sinh(self) -> Double

Computes the hyperbolic sine of the number.

Examples

let x = dd!(1).sinh();
let expected = dd!("1.1752011936438014568823818505956");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn cosh(self) -> Double

Computes the hyperbolic cosine of the number.

Examples

let x = dd!(1).cosh();
let expected = dd!("1.5430806348152437784779056207571");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn tanh(self) -> Double

Computes the hyperbolic tangent of the number.

Examples

let x = dd!(1).tanh();
let expected = dd!("0.76159415595576488811945828260479");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn sinh_cosh(self) -> (Double, Double)

Simultaneously computes the hyperbolic sine and cosine of the number.

This method is more efficient to run than sinh and cosh individually and is useful when both numbers are needed.

Examples

let (sin_h, cos_h) = dd!(1).sinh_cosh();
let esin = dd!("1.1752011936438014568823818505956");
let ecos = dd!("1.5430806348152437784779056207571");

let diff1 = (sin_h - esin).abs();
let diff2 = (cos_h - ecos).abs();

assert!(diff1 < dd!(1e-30));
assert!(diff2 < dd!(1e-30));

pub fn asinh(self) -> Double

Calculates the inverse hyperbolic sine of the number.

Examples

let x = dd!(1.5).asinh();
let expected = dd!("1.1947632172871093041119308285191");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn acosh(self) -> Double

Calculates the inverse hyperbolic cosine of the number.

Examples

let x = dd!(1.5).acosh();
let expected = dd!("0.96242365011920689499551782684874");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn atanh(self) -> Double

Calculates the inverse hyperbolic tangent of the number.

Examples

let x = dd!(0.5).atanh();
let expected = dd!("0.54930614433405484569762261846126");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

impl Double

pub fn abs(self) -> Double

Calculates the absolute value of the number.

Examples

assert!(dd!(3).abs() == dd!(3));
assert!(dd!(-3).abs() == dd!(3));

pub fn floor(self) -> Double

Returns the largest integer less than or equal to the number.

Examples

let f = dd!(3.99);
let g = dd!(3.0);

assert!(f.floor() == dd!(3));
assert!(g.floor() == dd!(3));

pub fn ceil(self) -> Double

Returns the smallest integer greater than or equal to the number.

Examples

let f = dd!(3.01);
let g = dd!(4.0);

assert!(f.ceil() == dd!(4));
assert!(g.ceil() == dd!(4));

pub fn round(self) -> Double

Returns the nearest integer to the number. Half-way cases are rounded away from 0.0.

Examples

let f = dd!(3.3);
let g = dd!(-3.3);

assert!(f.round() == dd!(3));
assert!(g.round() == dd!(-3));

pub fn trunc(self) -> Double

Returns the integer part of the number.

Examples

let f = dd!(3.3);
let g = dd!(-3.7);

assert!(f.trunc() == dd!(3));
assert!(g.trunc() == dd!(-3));

pub fn fract(self) -> Double

Returns the fractional part of the number.

Examples

let f = dd!(3.3);
let g = dd!(-3.7);

let fdiff = (f.fract() - dd!(0.3)).abs();
let gdiff = (g.fract() - dd!(-0.7)).abs();

assert!(fdiff < dd!(1e-30));
assert!(gdiff < dd!(1e-30));

pub fn signum(self) -> Double

Returns a number that represents the sign of the number.

1.0 if the number is positive, +0.0, or INFINITY

-1.0 if the number is negative, -0.0, or NEG_INFINITY

NAN if the number is NAN

Examples

assert!(dd!(3.5).signum() == Double::ONE);
assert!(Double::NEG_INFINITY.signum() == -Double::ONE);
assert!(Double::NAN.signum().is_nan());

impl Double

pub fn classify(self) -> FpCategory

Returns the floating point category of the number. If only one property is being tested, it's generally faster to use the specific predicate rather than this function.

Examples

use std::num::FpCategory;

let num = dd!(12.4);
let inf = Double::INFINITY;

assert!(num.classify() == FpCategory::Normal);
assert!(inf.classify() == FpCategory::Infinite);

pub fn is_normal(self) -> bool

Returns true if the number is neither zero, infinite, subnormal, or NaN.

Examples

let min = Double::MIN_POSITIVE;
let max = Double::MAX;
let lower = dd!(1e-308);
let zero = Double::ZERO;

assert!(min.is_normal());
assert!(max.is_normal());

assert!(!zero.is_normal());
assert!(!Double::NAN.is_normal());
assert!(!Double::INFINITY.is_normal());
// Values between `0` and `MIN_POSITIVE` are subnormal.
assert!(!lower.is_normal());

pub fn is_zero(self) -> bool

Returns true if the number is either positive or negative zero.

Examples

assert!(Double::ZERO.is_zero());
assert!(Double::NEG_ZERO.is_zero());
assert!(!Double::PI.is_zero());

pub fn is_sign_negative(self) -> bool

Returns true if the number is negative, including negative zero and infinity and NaN with a negative sign bit.

Examples

assert!(Double::NEG_ZERO.is_sign_negative());
assert!(Double::NEG_INFINITY.is_sign_negative());
assert!(dd!(-7.0).is_sign_negative());
assert!(!Double::ZERO.is_sign_negative());
assert!(!dd!(7.0).is_sign_negative());

pub fn is_sign_positive(self) -> bool

Returns true if the number is positive, including positive zero and infinity and NaN with a positive sign bit.

Examples

assert!(Double::ZERO.is_sign_positive());
assert!(Double::INFINITY.is_sign_positive());
assert!(dd!(7.0).is_sign_positive());
assert!(!Double::NEG_ZERO.is_sign_positive());
assert!(!dd!(-7.0).is_sign_positive());

pub fn is_nan(self) -> bool

Returns true if this value is NaN.

Examples

assert!(Double::NAN.is_nan());
assert!(!dd!(7.0).is_nan());

pub fn is_infinite(self) -> bool

Returns true if this number is positive or negative infinity.

Examples

assert!(Double::INFINITY.is_infinite());
assert!(Double::NEG_INFINITY.is_infinite());
assert!(!Double::NAN.is_infinite());
assert!(!dd!(7.0).is_infinite());

pub fn is_finite(self) -> bool

Returns true if this number is neither infinite nor NaN.

Examples

assert!(!Double::INFINITY.is_finite());
assert!(!Double::NEG_INFINITY.is_finite());
assert!(!Double::NAN.is_finite());
assert!(dd!(7.0).is_finite());

impl Double

pub fn new(a: f64, b: f64) -> Double

Creates a Double with the two arguments as the internal components.

This function is only useful in the simplest of cases, as it does not do normalization and therefore does not account for floating-point error in the first component (meaning the user has to). One of its primary functions is declaration of Double constants that have been pre-computed.

Examples

let dd = Double::new(0.0, 0.0);
assert!(dd.is_zero());

pub fn norm(a: f64, b: f64) -> Double

Creates a Double by normalizing the sum of two arguments.

This is a quick and efficient function, but it carries the restriction that the absolute value of a must be greater than or equal to the absolute value of b. If that cannot be guaranteed, it would be better to use the slightly slower but more robust from_add instead.

Examples

let dd = Double::norm(2.0, 1.0);
assert!(dd == dd!(3.0));