Struct qd::Quad

pub struct Quad(_, _, _, _);

A 256-bit floating-point number implemented as the unevaluated sum of four 64-bit floating-point numbers. Discarding the bits used for exponents, this makes for about 212 bits of mantissa accuracy, or around 63 decimal digits.

There are several ways to create a new Quad:

calling the new function

calling from with a primitive number or a string

calling parse on a string (or equivalently using from_str)

using the qd! macro

What kind of number you actually end up getting depends on the method called to get it. new will not normalize its result. This means that the arguments must be pre-normalized. from, parse, and qd! will both account for floating-point rounding error and produce normalized results.

The reason for these two different ways of going about creation is speed. If the number is already pre-computed to take normalization and error into account (as all of the constants in this library are), then new offers a way to avoid having to pay the efficiency cost of unnecessary normalization.

For the other methods, shortcuts can be taken if the input is a number and that number is dyadic (i.e., it can be represented in binary exactly, without rounding). In this case, from and qd! can also skip normalization and accounting for rounding, and they won't be much slower than new.

Parsing from strings or from numbers that are not dyadic cannot take these shortcuts. The results will be precise, but at the cost of speed.

See the module-level documentation for more information.

Implementations

impl Quad

pub const RADIX: u32

The radix or base of the internal representation of Quad. This is the same as the representation in the underlying f64.

pub const MANTISSA_DIGITS: u32

Number of significant digits in base 2.

pub const DIGITS: u32

Approximate number of significant digits in base 10.

pub const EPSILON: Quad

Machine epsilon value for Quad.

This is the difference between 1.0 and the next largest representable number.

pub const MIN: Quad

Smallest finite Quad value.

pub const MIN_POSITIVE: Quad

Smallest positive normal Quad value.

pub const MAX: Quad

Largest finite Quad value.

pub const MIN_EXP: i32

One greater than the minimum possible normal power of 2 exponent.

pub const MAX_EXP: i32

Maximum possible power of 2 exponent.

pub const MIN_10_EXP: i32

Minimum possible normal power of 10 exponent.

pub const MAX_10_EXP: i32

Maximum possible power of 10 exponent.

pub const NAN: Quad

Not a Number (NaN).

pub const INFINITY: Quad

Infinity ($\infin$).

pub const NEG_INFINITY: Quad

Negative infinity ($-\infin$).

pub const ZERO: Quad

Zero ($0$).

pub const NEG_ZERO: Quad

Negative zero ($-0$).

pub const ONE: Quad

One ($1$).

pub const NEG_ONE: Quad

Negative one ($-1$).

pub const PI: Quad

Archimedes' constant ($\pi$). The value of this constant is 3.1415926535897932384626433832795028841971693993751058209749446.

pub const TAU: Quad

The full circle constant ($\tau$), or $2\pi$. The value of this constant is 6.2831853071795864769252867665590057683943387987502116419498892.

pub const FRAC_PI_2: Quad

The constant $\pi/2$. The value of this constant is 1.5707963267948966192313216916397514420985846996875529104874723.

pub const FRAC_PI_3: Quad

The constant $\pi/3$. The value of this constant is 1.0471975511965977461542144610931676280657231331250352736583149.

pub const FRAC_PI_4: Quad

The constant $\pi/4$. The value of this constant is 0.78539816339744830961566084581987572104929234984377645524373615.

pub const FRAC_PI_6: Quad

The constant $\pi/6$. The value of this constant is 0.52359877559829887307710723054658381403286156656251763682915743.

pub const FRAC_PI_8: Quad

The constant $\pi/8$. The value of this constant is 0.39269908169872415480783042290993786052464617492188822762186807.

pub const FRAC_PI_16: Quad

The constant $\pi/16$. The value of this constant is 0.19634954084936207740391521145496893026232308746094411381093404.

pub const FRAC_3_PI_2: Quad

The constant $3\pi/2$. The value of this constant is 4.7123889803846898576939650749192543262957540990626587314624169.

pub const FRAC_3_PI_4: Quad

The constant $3\pi/4$. The value of this constant is 2.3561944901923449288469825374596271631478770495313293657312084.

pub const FRAC_5_PI_4: Quad

The constant $5\pi/4$. The value of this constant is 3.9269908169872415480783042290993786052464617492188822762186807.

pub const FRAC_7_PI_4: Quad

The constant $7\pi/4$. The value of this constant is 5.4977871437821381673096259207391300473450464489064351867061530.

pub const FRAC_1_PI: Quad

The constant $1/\pi$. The value of this constant is 0.31830988618379067153776752674502872406891929148091289749533469.

pub const FRAC_2_PI: Quad

The constant $2/\pi$. The value of this constant is 0.63661977236758134307553505349005744813783858296182579499066938.

pub const FRAC_2_SQRT_PI: Quad

The constant $2/\sqrt{\pi}$. The value of this constant is 1.1283791670955125738961589031215451716881012586579977136881714.

pub const SQRT_2: Quad

The constant $\sqrt{2}$. The value of this constant is 1.4142135623730950488016887242096980785696718753769480731766797.

pub const FRAC_1_SQRT_2: Quad

The constant $1/\sqrt{2}$. The value of this constant is 0.70710678118654752440084436210484903928483593768847403658833987.

pub const E: Quad

Euler's number ($e$). The value of this constant is 2.7182818284590452353602874713526624977572470936999595749669676.

pub const LOG2_10: Quad

The constant $\log_2 10$. The value of this constant is 3.3219280948873623478703194294893901758648313930245806120547564.

pub const LOG2_E: Quad

The constant $\log_2 e$. The value of this constant is 1.4426950408889634073599246810018921374266459541529859341354494.

pub const LOG10_2: Quad

The constant $\log_{10} 2$. The value of this constant is 0.30102999566398119521373889472449302676818988146210854131042746.

pub const LOG10_E: Quad

The constant $\log_{10} e$. The value of this constant is 0.43429448190325182765112891891660508229439700580366656611445378.

pub const LN_2: Quad

The constant $\ln 2$. The value of this constant is 0.69314718055994530941723212145817656807550013436025525412068001.

pub const LN_10: Quad

The constant $\ln 10$. The value of this constant is 2.3025850929940456840179914546843642076011014886287729760333279.

impl Quad

pub fn ldexp(self, n: i32) -> Quad

Calculates $x\sdot2^n$, where $x$ is self and $n$ is the argument.

Though this is not an everyday operation, it is often used in more advanced mathematical calculations (including several within this library). Therefore an implementation that is much more efficient than calculating it through multiplication and powi is offered despite it not being part of the f64 API.

Examples

let x = qd!(3);
assert!(x.ldexp(5) == qd!(96)); // 3 * 2^5

pub fn sqr(self) -> Quad

Calculates the square of $x$, $x^2$, where $x$ is self.

This method takes advantage of optimizations in multiplication that are available when the two numbers being multiplied are the same, so it is more efficient than bare multiplication.

Examples

let x = qd!(3);
assert!(x.sqr() == x * x); // The left side is faster though

pub fn sqrt(self) -> Quad

Calculates the square root $x$, $\sqrt{x}$, where $x$ is self.

Examples

let x = qd!(2).sqrt();
let expected = qd!("1.414213562373095048801688724209698078569671875376948073176679738");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn nroot(self, n: i32) -> Quad

Calculates the $n$th root of $x$, $\sqrt[n]{x}$, where $x$ is self and $n$ is the argument.

Examples

let x = qd!(2).nroot(4);
let expected = qd!("1.189207115002721066717499970560475915292972092463817413019002225");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn cbrt(self) -> Quad

Calculates the cube root of $x$, $\sqrt[3]{x}$, where $x$ is self.

Examples

let x = qd!(2).cbrt();
let expected = qd!("1.259921049894873164767210607278228350570251464701507980081975112");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn powi(self, n: i32) -> Quad

Calculates $x$ raised to the integer power $n$, $x^n$, where $x$ is self and $n$ is the argument.

This function correctly handles the special inputs defined in IEEE 754. In particular:

x.powi(0) is $1$ for any x (including 0, NaN, or inf)

x.powi(n) is $\pm\infin$ for x = $\pm0$ and any odd negative n

x.powi(n) is $+\infin$ for x = $\pm0$ and any even negative n

x.powi(n) is $\pm0$ for x = $\pm0$ and any odd positive n

x.powi(n) is $+0$ for x = $\pm0$ and any even positive n

Examples

let x = qd!(3);
assert!(x.powi(3) == qd!(27));

pub fn powf(self, n: Quad) -> Quad

Calculates $x$ raised to the Quad power $n$, $x^n$, where $x$ is self and $n$ is the argument.

In general, $x^n$ is equal to $e^{n \ln x}$. This precludes raising a negative Double to a fractional or irrational power because $\ln x$ is undefined when $x$ is negative. In that case, this function returns NAN.

It's actually more complex than that; if the exponent can be expressed as a fraction with an odd denominator, then there is an answer ($\sqrt[3]{x}$, which is defined for negative $x$, is the same as a $x^\frac{1}{3}$). Therefore, something like qd!(-4).powf(qd!(0.2)) should work, as 0.2 is a fraction with an odd denominator ($\frac{1}{5}$), and the entire expression is the same as $\sqrt[5]{-4}$, which is a real thing. However, it's impossible in general to tell whether a number is a fraction while using floating-point numbers, so no attempt is made to make this work. If you need a fifth root of 4, use qd!(-4).nroot(5).

Examples

let x = qd!(3).powf(qd!(3.3));
let expected = qd!("37.54050759852955219310186595463382927684873090166843452920390518");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn recip(self) -> Quad

Calculates the reciprocal of $x$, $\frac{1}{x}$, where $x$ is self.

Examples

let x = Quad::PI.recip();
let expected = qd!("0.3183098861837906715377675267450287240689192914809128974953346881");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

impl Quad

pub fn rem_euclid(self, rhs: Quad) -> Quad

Calculates the least non-negative remainder of $x \div y$, where $x$ is self and $y$ is the argument. This is $x \mod y$ using the Euclidean definition of the modulo operation.

In principle, the returned value $r$ will satisfy $0 \le r < |y|$; it will be non-negative and less than the absolute value of the argument. However, in extreme cases, $r$ may equal $|y|$ because of floating-point rounding error.

Examples

let x = qd!(21);
let y = qd!(4);

assert!(x.rem_euclid(y) == qd!(1));
assert!((-x).rem_euclid(y) == qd!(3));
assert!(x.rem_euclid(-y) == qd!(1));
assert!((-x).rem_euclid(-y) == qd!(3));

pub fn div_euclid(self, rhs: Quad) -> Quad

Calculates the quotient in the Euclidean division $x \div y$, where $x$ is self and $y$ is the argument.

The Euclidean quotient is x / y rounded to the integer n such that x >= n * y. That makes it the counterpart to rem_euclid in the same way that regular division is the counterpart to the % operator; the answer n that it computes is the integer such that x = n * y + x.rem_euclid(y).

Examples

let x = qd!(21);
let y = qd!(4);

assert!(x.div_euclid(y) == qd!(5));
assert!((-x).div_euclid(y) == qd!(-6));
assert!(x.div_euclid(-y) == qd!(-5));
assert!((-x).div_euclid(-y) == qd!(6));

impl Quad

pub fn sinh_cosh(self) -> (Quad, Quad)

Simultaneously computes the hyperbolic sine and cosine of $x$, $\sinh x$ and $\cosh x$, where $x$ is self.

The domain of this function is $(-\infin, \infin)$, and the range is $(-\infin, \infin)$ for the first component of the answer (the hyperbolic sine) and $[1, \infin)$ for the second (the hyperbolic cosine).

This method is more efficient to run than sinh and cosh individually and is useful when both numbers are needed.

Examples

let (sin_h, cos_h) = qd!(1).sinh_cosh();
let esin = qd!("1.175201193643801456882381850595600815155717981334095870229565413");
let ecos = qd!("1.543080634815243778477905620757061682601529112365863704737402215");

let diff1 = (sin_h - esin).abs();
let diff2 = (cos_h - ecos).abs();

assert!(diff1 < qd!(1e-60));
assert!(diff2 < qd!(1e-60));

pub fn sinh(self) -> Quad

Computes the hyperbolic sine of $x$, $\sinh x$, where $x$ is self.

The domain and range of this function are both $(-\infin, \infin)$.

Examples

let x = qd!(1).sinh();
let expected = qd!("1.175201193643801456882381850595600815155717981334095870229565413");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn cosh(self) -> Quad

Computes the hyperbolic cosine of $x$, $\cosh x$, where $x$ is self.

The domain of this function is $(-\infin, \infin)$, and the range is $[1, \infin)$.

Examples

let x = qd!(1).cosh();
let expected = qd!("1.543080634815243778477905620757061682601529112365863704737402215");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn tanh(self) -> Quad

Computes the hyperbolic tangent of $x$, $\tanh x$, where $x$ is self.

The domain of this function is $(-\infin, \infin)$, and the range is $(-1, 1)$.

Examples

let x = qd!(1).tanh();
let expected = qd!("0.7615941559557648881194582826047935904127685972579365515968105001");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn asinh(self) -> Quad

Calculates the inverse hyperbolic sine of $x$, $\sinh^{-1} x$, where $x$ is self.

The domain and range of this function are both $(-\infin, \infin)$.

Examples

let x = qd!(1.5).asinh();
let expected = qd!("1.194763217287109304111930828519090523536162075153005429270680299");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn acosh(self) -> Quad

Calculates the inverse hyperbolic cosine of $x$, $\cosh^{-1} x$, where $x$ is self.

The domain of the function is $[1, \infin)$ and the range is $[0, \infin)$. Any argument outside the range will result in NAN.

Examples

let x = qd!(1.5).acosh();
let expected = qd!("0.9624236501192068949955178268487368462703686687713210393220363377");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-30));

pub fn atanh(self) -> Quad

Calculates the inverse hyperbolic tangent of $x$, $\tanh^{-1} x$, where $x$ is self.

The domain of the function is $(-1, 1)$ and the range is $(-\infin, \infin)$. Any argument whose absolute value is greater than or equal to 1 will result in NAN.

Examples

let x = qd!(0.5).atanh();
let expected = qd!("0.5493061443340548456976226184612628523237452789113747258673471668");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

impl Quad

pub fn abs(self) -> Quad

Calculates the absolute value of $x$, $|x|$, where $x$ is self. The absolute value of $x$ is simply the same value as $x$, but with the opposite sign if $x$ is negative.

Examples

assert!(qd!(3).abs() == qd!(3));
assert!(qd!(-3).abs() == qd!(3));

pub fn floor(self) -> Quad

Calculates the floor of $x$, $\lfloor{x}\rfloor$, where $x$ is self.

The floor of $x$ is the largest integer value less than or equal to $x$. This means that the floor of a negative number will have an absolute value greater than that of the number itself.

Examples

let f = qd!(3.99);
let g = qd!(3.0);

assert!(f.floor() == qd!(3));
assert!(g.floor() == qd!(3));

pub fn ceil(self) -> Quad

Calculates the ceiling of $x$, $\lceil{x}\rceil$, where $x$ is self.

The ceiling of $x$ is the smallest integer value greater than or equal to $x$. This means that the ceiling of a negative number will have an absolute value the same as (not greater than) that of the number itself.

Examples

let f = qd!(3.01);
let g = qd!(4.0);

assert!(f.ceil() == qd!(4));
assert!(g.ceil() == qd!(4));

pub fn round(self) -> Quad

Calculates the rounded value of $x$, where $x$ is self.

The rounded value is the nearest integer to $x$. Halfway cases (i.e., numbers with a fractional portion of 0.5) are rounded away from 0, per the behavior of f64's round function.

Examples

let f = qd!(3.3);
let g = qd!(-3.3);

assert!(f.round() == qd!(3));
assert!(g.round() == qd!(-3));

pub fn trunc(self) -> Quad

Returns the integer part of self. This integer part will be of the same sign as the original number.

Examples

let f = qd!(3.3);
let g = qd!(-3.7);

assert!(f.trunc() == qd!(3));
assert!(g.trunc() == qd!(-3));

pub fn fract(self) -> Quad

Returns the fractional part of the self. This fractional part will be of the same sign as the original number.

Examples

let f = qd!(3.3);
let g = qd!(-3.7);

let fdiff = (f.fract() - qd!(0.3)).abs();
let gdiff = (g.fract() - qd!(-0.7)).abs();

assert!(fdiff < qd!(1e-60));
assert!(gdiff < qd!(1e-60));

pub fn signum(self) -> Quad

Returns a number that represents the sign of self.

ONE if self is positive, including +0.0 and INFINITY

NEG_ONE if self is negative, including -0.0 and NEG_INFINITY

NAN if self is NAN

Examples

assert!(qd!(3.5).signum() == Quad::ONE);
assert!(Quad::NEG_INFINITY.signum() == Quad::NEG_ONE);
assert!(Quad::NAN.signum().is_nan());

pub fn classify(self) -> FpCategory

Returns self's floating point category.

The possible return values are the members of FpCategory, as follows:

FpCategory::Zero if the number is $\pm0$;

FpCategory::Infinite if the number is $\pm\infin$;

FpCategory::Nan if the number is not a number;

FpCategory::Subnormal if the number is $\pm$MIN_POSITIVE (numbers this small can be represented, but they lose some accuracy);

FpCategory::Normal if the number is anything else.

A Quad can also register as FpCategory::Subnormal if it has a small enough negative exponent that one of the other components of the number is a subnormal number itself. This will typically happen around 1e-260 or so.

Examples

use std::num::FpCategory;

let num = qd!(12.4);
let inf = Quad::INFINITY;

assert!(num.classify() == FpCategory::Normal);
assert!(inf.classify() == FpCategory::Infinite);

pub fn is_normal(self) -> bool

Returns true if self is neither zero, infinite, subnormal, or NaN.

Examples

let min = Quad::MIN_POSITIVE;
let max = Quad::MAX;
let lower = qd!(1e-308);
let zero = Quad::ZERO;

assert!(min.is_normal());
assert!(max.is_normal());

assert!(!zero.is_normal());
assert!(!Quad::NAN.is_normal());
assert!(!Quad::INFINITY.is_normal());
// Values between `0` and `MIN_POSITIVE` are subnormal.
assert!(!lower.is_normal());

pub fn is_zero(self) -> bool

Returns true if self is either positive or negative zero.

Examples

assert!(Quad::ZERO.is_zero());
assert!(Quad::NEG_ZERO.is_zero());
assert!(!Quad::PI.is_zero());

pub fn is_sign_negative(self) -> bool

Returns true if self is negative, including negative zero, negative infinity, and NaN with a negative sign bit.

Examples

assert!(Quad::NEG_ZERO.is_sign_negative());
assert!(Quad::NEG_INFINITY.is_sign_negative());
assert!(qd!(-7.0).is_sign_negative());
assert!(!Quad::ZERO.is_sign_negative());
assert!(!qd!(7.0).is_sign_negative());

pub fn is_sign_positive(self) -> bool

Returns true if self is positive, including positive zero, positive infinity and NaN with a positive sign bit.

Examples

assert!(Quad::ZERO.is_sign_positive());
assert!(Quad::INFINITY.is_sign_positive());
assert!(qd!(7.0).is_sign_positive());
assert!(!Quad::NEG_ZERO.is_sign_positive());
assert!(!qd!(-7.0).is_sign_positive());

pub fn is_nan(self) -> bool

Returns true if self is NaN.

This is the proper way to test for NaN because it cannot be done with an equality test (since NaN is not equal to itself).

Examples

assert!(Quad::NAN.is_nan());
assert!(!qd!(7.0).is_nan());

pub fn is_infinite(self) -> bool

Returns true if self is positive or negative infinity.

Examples

assert!(Quad::INFINITY.is_infinite());
assert!(Quad::NEG_INFINITY.is_infinite());
assert!(!Quad::NAN.is_infinite());
assert!(!qd!(7.0).is_infinite());

pub fn is_finite(self) -> bool

Returns true if self is neither infinite nor NaN.

Examples

assert!(!Quad::INFINITY.is_finite());
assert!(!Quad::NEG_INFINITY.is_finite());
assert!(!Quad::NAN.is_finite());
assert!(qd!(7.0).is_finite());

pub fn is_subnormal(self) -> bool

Returns true if self has an absolute value of less than MIN_POSITIVE.

Numbers this small can be represented by floating point numbers, but they are not as accurate. This inaccuracy is inherent in the IEEE-754 format for 64-bit numbers; making a double-double out of an inaccurate number means the double-double is also going to be inaccurate.

Examples

assert!(!Quad::PI.is_subnormal());
assert!(qd!(1e-308).is_subnormal());

impl Quad

pub fn exp(self) -> Quad

Computes the exponential function, $e^x$, where $x$ is self.

The result of this function grows rapidly. Once x exceeds 708, the result is too large to represent with a Quad; at that point the function begins to return INFINITY. The limit on the low end is less due to the fact that the second, third, and fourth components need to fit in an f64 rather than the first, along with extra bits used in argument reduction; this function begins to return 0 at -460.

As x grows this function does lose a bit of precision. It's precise to at least 60 digits up to values of $-140 \le x \le 150$, and from then until the limits, it's precise to at least 59 digits.

Examples

let x = qd!(2.3).exp();
let expected = qd!("9.974182454814720739957615156908858001478701193684029563691421917");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-59));

pub fn expm1(self) -> Quad

Computes the exponential function of $x$ minus 1, $e^x - 1$, where $x$ is self.

While this function literally calculates the value returned by exp minus 1, it does this directly (rather than computing exp directly and then subtracting 1 from the answer). This is useful in the not-infrequent case where $x$ is very close to 0 and you have to subtract something near 1 from the answer (another example is x.exp() - x.cos() — for very small values of x, x.cos() is also very near 1). Since x.exp() is very close to 1 when x is very close to 0, this is subtracting something very close to 1 from something else very close to 1.

When this happens, subtracting 1 from exp suffers from catastrophic cancellation, a condition in which subtracting two numbers that are very close to each other can result in a huge loss of accuracy. Since expm1 does not perform this subtraction but instead computes ex - 1 directly, it does not suffer from this phenomenon.

This function will work over the same range as exp, but x.expm1() + 1 has no advantage over x.exp(). It's for that one very particular purpose.

Examples

// A value very close to 0.
let v_small = qd!(1e-58);
// The actual value of exp(v_small) - 1. The last digit is on the order of 1e-125.
let expected =
    qd!("1.0000000000000000000000000000000000000000000000000000000000500000004e-58");

// First with exp, subtracting one from the result. The resulting precision loses at
// least 14 digits.
let exp = v_small.exp() - Quad::ONE;
let eps = (exp - expected).abs();
assert!(eps > qd!(1e-108));

// Next with expm1. Since this is calculated without subtracting one from anything,
// the catastrophic cancellation does not occur and full precision is maintained.
let expm1 = v_small.expm1();
let epsm1 = (expm1 - expected).abs();
assert!(epsm1 < qd!(1e-122));

pub fn ln(self) -> Quad

Calculates the base-$e$ (natural) logarithm of $x$, $\ln x$ or $\log_e x$, where $x$ is self.

This calculation relies upon the exp calculation, in the opposite direction. A large positive logarithm, for example, will require the calculation of a large negative exponential.

For the same reasons that negative values of exp are limited to -470, the accurate results of this function are limited to the number whose logarithm is 460, which is around 2 × 10200. Take care with this; unlike in exp, INFINITY is not returned. In that function, exceeding the maximum refers to actually overflowing an f64, which is appropriate to call INFINITY; here, it means 470.

Examples

let x = qd!(7).ln();
let expected = qd!("1.945910149055313305105352743443179729637084729581861188459390150");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn ln1p(self) -> Quad

Calculates the base-$e$ (natural) logarithm of 1 plus $x$, $\ln (1 + x)$ or $\log_e (1 + x)$, where $x$ is self.

This is the inverse of expm1 and arises from the same sorts of concerns. It isn't unusual to want to take logarithms of numbers very near 1, as the logarithm approaches 0 at that point. However, with finite-precision mathematics, the 1 itself severely limits the precision possible; the number 1.000000000000001 has 16 digits of precision, but most of them are taken up by placeholder zeros when we would prefer to have that precision available after the final 1.

ln1p allows that by letting the user pass in a number near 0 and having the algorithm add 1 to it internally, without causing the loss of precision. For example, the same 16-digit number above could be passed into ln1p as 0.000000000000001, a number with one digit of precision, leaving 15 more digits of precision availble after that final 1.

The algorithm for logarithms close to 1 is slower than that for the general logarithm, so this function delegates to ln if it can be done without losing precision. There is no advantage to using ln1p over ln except for computing logarithms of numbers very close to 1.

Examples

// A very small number
let v_small = qd!(1e-150);
// The actual value for the base-e logarithm of 1 + 1e-150
let expected =
    qd!("9.9999999999999999999999999999999999999999999999999999999999999999996e-151");

// First with ln, computing the logarithm of 1 + 1e-150. This loses at least 14
// digits of precision because of the precision-damping effect of adding the one
// before the computation.
let ln = (Quad::ONE + v_small).ln();
let lneps = (ln - expected).abs();
assert!(lneps > qd!(1e-200));

// Now with ln1p, computing the same logarithm. This does not suffer from the same
// effect from the 1, and the answer retains full precision.
let ln1p = v_small.ln1p();
let ln1peps = (ln1p - expected).abs();
assert!(ln1peps < qd!(1e-214));

pub fn log10(self) -> Quad

Calculates the base-10 logarithm of $x$, $\log_{10} x$, where $x$ is self.

As with ln, this has an upper usable range less than the size of the numbers themselves. In this case, that upper limit is around 1e200. Over this number, the output is not reliable, but it does not return [INFINITY`] because the number 200 is so plainly not infinite.

Examples

let x = Quad::E.log10();
let expected = qd!("0.4342944819032518276511289189166050822943970058036665661144537832");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn log2(self) -> Quad

Calculates the base-2 logarithm of $x$, $\log_2 x$, where $x$ is self.

Since 2 is smaller than $e$, this function is constrained even more than ln. It will start returning NEG_INFINITY at around 1e-213 and will start to fail on the positive side at around 2.6e180.

Examples

let x = qd!(10).log2();
let expected = qd!("3.321928094887362347870319429489390175864831393024580612054756396");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-59));

pub fn log(self, b: Quad) -> Quad

Calculates the base-$b$ logarithm of $x$, $\log_b x$, where $x$ is self and $b$ is the argument.

This function will have limits at extreme arguments like the other logarithm functions. The difference is that those limits will depend on the base argument.

If the goal is to calculate the base-$e$, base-2, or base-10 logarithms of $x$, the specialized functions for those purposes (ln, log2, and log10 respectively) will be more efficient.

Examples

let x = qd!(10).log(qd!(7.3));
let expected = qd!("1.158315209978887965104764376269736420106652944692834002126233653");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-59));

impl Quad

pub fn sin_cos(self) -> (Quad, Quad)

Simultaneously computes the sine and cosine of $x$, $\sin x$ and $\cos x$, where $x$ is self. This is more efficient than calling the separate sin and cos functions if you need both numbers.

Examples

let x = Quad::PI / qd!(4);
let (sin_x, cos_x) = x.sin_cos();

let diff_sin = (sin_x - x.sin()).abs();
let diff_cos = (cos_x - x.cos()).abs();

assert!(diff_sin < qd!(1e-60));
assert!(diff_cos < qd!(1e-60));

pub fn sin(self) -> Quad

Computes the sine of $x$, $\sin x$, where $x$ is self.

The domain of this function is $(-\infin, \infin)$, and the range is $[-1, 1]$.

Examples

let x = qd!(1).sin();
let expected = qd!("0.8414709848078965066525023216302989996225630607983710656727517099884");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn cos(self) -> Quad

Computes the cosine of $x$, $\cos x$, where $x$ is self.

The domain of this function is $(-\infin, \infin)$, and the range is $[-1, 1]$.

Examples

let x = qd!(1).cos();
let expected = qd!("0.5403023058681397174009366074429766037323104206179222276700972553787");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn tan(self) -> Quad

Computes the tangent of $x$, $\tan x$, where $x$ is self.

The domain and range of this function are both $(-\infin, \infin)$.

Examples

let x = qd!(1).tan();
let expected = qd!("1.557407724654902230506974807458360173087250772381520038383946606");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn atan2(self, other: Quad) -> Quad

Computes the 2-argument inverse tangent of $x$ and $y$, $\arctan(y, x)$, where $x$ is the argument and $y$ is self.

The single-argument atan function always returns values in either the first (0 to $\pi$/2) or fourth (0 to $-\pi$/2) quadrants. However, first-quadrant results repeat themselves in the third quadrant, and fourth-quadrant results repeat themselves in the second. For example, $\tan \frac{\pi}{4}$ is 1, but so is $\tan \frac{3\pi}{4}$. Single-argument atan cannot distinguish between these two possibilities, so it always returns the one in the range $[-\frac{\pi}{2}, \frac{\pi}{2}]$.

The double-argument atan2 can return either, depending on the arguments. It essentially returns the angle between the positive x-axis and the point $(x, y)$, where $y$ is self and $x$ is the argument. Therefore dd!(1).atan2(dd!(1)) returns $\frac{\pi}{4}$ (first quadrant), but flipping both signs to dd!(-1).atan2(dd!(-1)) gives the $-\frac{3\pi}{4}$ result (third quadrant).

This function extends the range of the result to $[-\pi, \pi]$.

Because this function deals with angles around the origin and Cartesian coordinates, it's very useful for converting between Cartesian and polar coordinates.

Examples

let pi = Quad::PI;

// -π/4 radians (45 degrees clockwise)
let x1 = qd!(3);
let y1 = qd!(-3);
let expected1 = -pi / qd!(4);

// 3π/4 radians (135 degrees counter-clockwise)
let x2 = qd!(-3);
let y2 = qd!(3);
let expected2 = Quad::from(0.75) * pi;

let diff1 = (y1.atan2(x1) - expected1).abs();
let diff2 = (y2.atan2(x2) - expected2).abs();

assert!(diff1 < qd!(1e-60));
assert!(diff2 < qd!(1e-60));

pub fn asin(self) -> Quad

Computes the inverse sine of $x$, $\sin^{-1} x$, where $x$ is self.

The domain of this function is $[-1, 1]$ while the range is $[-\frac{\pi}{2}, \frac{\pi}{2}]$. Arguments outside of this domain will result in NAN.

Examples

let x = qd!(1).asin();
let expected = Quad::PI / qd!(2);  // π/2

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn acos(self) -> Quad

Computes the inverse cosine of $x$, $\cos^{-1} x$, where $x$ is self.

The domain of this function is $[-1, 1]$ and the range is $[0, \pi]$. Arguments outside of the domain will result in NAN.

Examples

let x = qd!(1).acos();
let expected = qd!(0);

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn atan(self) -> Quad

Computes the inverse tangent of $x$, $\tan^{-1} x$, where $x$ is self.

This is the single-argument version of this function (atan2 is the two-argument inverse tangent). Though inverse tangents have multiple answers, this function will always return the one with the lowest absolute value. x.atan() is the same as x.atan2(Quad::ONE).

The domain of this function is $(-\infin, \infin)$ and the range is $[-\frac{\pi}{2}, \frac{\pi}{2}]$.

Examples

let x = qd!(1).atan();
let expected = Quad::PI / qd!(4);  // π/4

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

impl Quad

pub const fn new(a: f64, b: f64, c: f64, d: f64) -> Quad

Creates a Quad with the four arguments as the internal components.

Be sure you know what you're doing if you use this function. It does not normalize its components, meaning that if they aren't already normalized by the caller, this number will not work the way one would expect (it'll fail equality tests that it should pass, it may be classified incorrectly, etc.).

This function is primarily for creating constants where the normalization is obviously unnecessary. For example, if a Quad version of the number 10 is needed, Quad::new(10.0, 0.0, 0.0, 0.0) is a good way to do it in order to save the cost of the normalization that is obviously not needed.

Examples

let q = Quad::new(0.0, 0.0, 0.0, 0.0);
assert!(q.is_zero());