Struct qd::Double

pub struct Double(_, _);

A 128-bit floating-point number implemented as the unevaluated sum of two 64-bit floating-point numbers. Discarding the bits used for exponents, this makes for about 106 bits of mantissa accuracy, or around 31 decimal digits.

There are several ways to create a new Double:

calling the new function

calling from with a primitive number (except for u128 and i128) or a string

calling parse on a string (or equivalently using from_str)

using the dd! macro

What kind of number you actually end up getting depends on the method called to get it. new will not normalize its result. This means that the arguments must be pre-normalized. from, parse, and dd! will both account for floating-point rounding error and produce normalized results.

The reason for these two different ways of going about creation is speed. If the number is already pre-computed to take normalization and error into account (as all of the constants in this library are), then new offers a way to avoid having to pay the efficiency cost of unnecessary normalization.

For the other methods, shortcuts can be taken if the input is a number and that number is dyadic (i.e., it can be represented in binary exactly, without rounding). In this case, from and dd! can also skip normalization and accounting for rounding, and they won't be much slower than new.

Parsing from strings or from numbers that are not dyadic cannot take these shortcuts. The results will be precise, but at the cost of speed.

See the module-level documentation for more information.

Implementations

impl Double

pub const RADIX: u32

The radix or base of the internal representation of Double. This is the same as the representation in the underlying f64.

pub const MANTISSA_DIGITS: u32

Number of significant digits in base 2.

pub const DIGITS: u32

Approximate number of significant digits in base 10.

pub const EPSILON: Double

Machine epsilon value for Double.

This is the difference between 1.0 and the next largest representable number.

pub const MIN: Double

Smallest finite Double value.

pub const MIN_POSITIVE: Double

Smallest positive normal Double value.

pub const MAX: Double

Largest finite Double value.

pub const MIN_EXP: i32

One greater than the minimum possible normal power of 2 exponent.

pub const MAX_EXP: i32

Maximum possible power of 2 exponent.

pub const MIN_10_EXP: i32

Minimum possible normal power of 10 exponent.

pub const MAX_10_EXP: i32

Maximum possible power of 10 exponent.

pub const NAN: Double

Not a Number (NaN).

pub const INFINITY: Double

Infinity ($\infin$).

pub const NEG_INFINITY: Double

Negative infinity ($-\infin$).

pub const ZERO: Double

Zero ($0$).

pub const NEG_ZERO: Double

Negative zero ($-0$).

pub const ONE: Double

One ($1$).

pub const NEG_ONE: Double

Negative one ($-1$).

pub const PI: Double

Archimedes' constant ($\pi$). The value of this constant is 3.141592653589793238462643383280.

pub const TAU: Double

The full circle constant ($\tau$), or $2\pi$. The value of this constant is 6.283185307179586476925286766559.

pub const FRAC_PI_2: Double

The constant $\pi/2$. The value of this constant is 1.570796326794896619231321691640.

pub const FRAC_PI_3: Double

The constant $\pi/3$. The value of this constant is 1.047197551196597746154214461093.

pub const FRAC_PI_4: Double

The constant $\pi/4$. The value of this constant is 0.7853981633974483096156608458199.

pub const FRAC_PI_6: Double

The constant $\pi/6$. The value of this constant is 0.5235987755982988730771072305466.

pub const FRAC_PI_8: Double

The constant $\pi/8$. The value of this constant is 0.3926990816987241548078304229099.

pub const FRAC_PI_16: Double

The constant $\pi/16$. The value of this constant is 0.1963495408493620774039152114550.

pub const FRAC_3_PI_2: Double

The constant $3\pi/2$. The value of this constant is 4.712388980384689857693965074919.

pub const FRAC_3_PI_4: Double

The constant $3\pi/4$. The value of this constant is 2.356194490192344928846982537460.

pub const FRAC_5_PI_4: Double

The constant $5\pi/4$. The value of this constant is 3.926990816987241548078304229099.

pub const FRAC_7_PI_4: Double

The constant $7\pi/4$. The value of this constant is 5.497787143782138167309625920739.

pub const FRAC_1_PI: Double

The constant $1/\pi$. The value of this constant is 0.3183098861837906715377675267450.

pub const FRAC_2_PI: Double

The constant $2/\pi$. The value of this constant is 0.6366197723675813430755350534901.

pub const FRAC_2_SQRT_PI: Double

The constant $2/\sqrt{\pi}$. The value of this constant is 1.128379167095512573896158903122.

pub const SQRT_2: Double

The constant $\sqrt{2}$. The value of this constant is 1.414213562373095048801688724210.

pub const FRAC_1_SQRT_2: Double

The constant $1/\sqrt{2}$. The value of this constant is 0.7071067811865475244008443621048.

pub const E: Double

Euler's number ($e$). The value of this constant is 2.718281828459045235360287471353.

pub const LOG2_10: Double

The constant $\log_2 10$. The value of this constant is 3.321928094887362347870319429489.

pub const LOG2_E: Double

The constant $\log_2 e$. The value of this constant is 1.442695040888963407359924681002.

pub const LOG10_2: Double

The constant $\log_{10} 2$. The value of this constant is 0.3010299956639811952137388947245.

pub const LOG10_E: Double

The constant $\log_{10} e$. The value of this constant is 0.4342944819032518276511289189166.

pub const LN_2: Double

The constant $\ln 2$. The value of this constant is 0.6931471805599453094172321214582.

pub const LN_10: Double

The constant $\ln 10$. The value of this constant is 2.302585092994045684017991454684.

impl Double

pub fn ldexp(self, n: i32) -> Double

Calculates $x\sdot2^n$, where $x$ is self and $n$ is the argument.

Though this is not an everyday operation, it is often used in more advanced mathematical calculations (including several within this library). Therefore an implementation that is much more efficient than calculating it through multiplication and powi is offered despite it not being part of the f64 API.

Examples

let x = dd!(5);
assert!(x.ldexp(3) == dd!(40)); // 5 * 2^3

pub fn agm(self, other: Double) -> Double

Calculates the arithmetic-geometric mean of $x$ and $y$ ($M(x, y)$ or $\text{agm}(x, y)$), where $x$ and $y$ are self and the argument.

The AGM is an iterative calculation. $x$ and $y$ are assigned as the two inputs to the first iteration.

$$a_0 = x \newline g_0 = y$$

Then the iterations are performed as two interdependent sequences, one calculating the arithmetic mean and one the geometric mean. The results from each iteration are fed back into the next iteration as inputs.

$$a_{n + 1} = \frac{a_n + g_n}{2}$$ $$g_{n + 1} = \sqrt{a_{n}g_{n}}$$

These numbers both converge towards the same number, and once they reach that number, it's returned as the AGM.

These sequences converge very quickly, doubling the accuracy with each iteration. The AGM can be used in fast algorithms for transcendental functions and for computing some mathematical constants (most notably $\pi$). The speed advantage from its fast convergence is realized at around 400 decimal digits, so it isn't currently used for these purposes in this library.

Examples

let x = Double::E.agm(Double::PI);
let expected = dd!("2.9261085515723046966658957101705574");

let delta = (x - expected).abs();
assert!(delta < dd!(1e-30));

pub fn sqr(self) -> Double

Calculates the square of $x$, $x^2$, where $x$ is self.

This method takes advantage of optimizations in multiplication that are available when the two numbers being multiplied are the same, so it is more efficient than bare multiplication.

Examples

let x = dd!(3);
assert!(x.sqr() == x * x); // The left side is faster though

pub fn sqrt(self) -> Double

Calculates the square root $x$, $\sqrt{x}$, where $x$ is self.

Examples

let x = dd!(2).sqrt();
let diff = (x - Double::SQRT_2).abs();
assert!(diff < dd!(1e-30));

pub fn nroot(self, n: i32) -> Double

Calculates the $n$th root of $x$, $\sqrt[n]{x}$, where $x$ is self and $n$ is the argument.

Examples

let x = dd!(2).nroot(4);
let expected = dd!("1.1892071150027210667174999705605");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn cbrt(self) -> Double

Calculates the cube root of $x$, $\sqrt[3]{x}$, where $x$ is self.

Examples

let x = dd!(2).cbrt();
let expected = dd!("1.2599210498948731647672106072782");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn powi(self, n: i32) -> Double

Calculates $x$ raised to the integer power $n$, $x^n$, where $x$ is self and $n$ is the argument.

This function correctly handles the special inputs defined in IEEE 754. In particular:

x.powi(0) is $1$ for any x (including 0, NaN, or inf)

x.powi(n) is $\pm\infin$ for x = $\pm0$ and any odd negative n

x.powi(n) is $+\infin$ for x = $\pm0$ and any even negative n

x.powi(n) is $\pm0$ for x = $\pm0$ and any odd positive n

x.powi(n) is $+0$ for x = $\pm0$ and any even positive n

Examples

let x = dd!(3);
assert!(x.powi(3) == dd!(27));

pub fn powf(self, n: Double) -> Double

Calculates $x$ raised to the Double power $n$, $x^n$, where $x$ is self and $n$ is the argument.

In general, $x^n$ is equal to $e^{n \ln x}$. This precludes raising a negative Double to a fractional or irrational power because $\ln x$ is undefined when $x$ is negative. In that case, this function returns NAN.

It's actually more complex than that; if the exponent can be expressed as a fraction with an odd denominator, then there is an answer ($\sqrt[3]{x}$, which is defined for negative $x$, is the same as a $x^\frac{1}{3}$). Therefore, something like dd!(-4).powf(dd!(0.2)) should work, as 0.2 is a fraction with an odd denominator ($\frac{1}{5}$), and the entire expression is the same as $\sqrt[5]{-4}$, which is a real thing. However, it's impossible in general to tell whether a number is a fraction while using floating-point numbers, so no attempt is made to make this work. If you need a fifth root of 4, use dd!(-4).nroot(5).

Examples

let x = dd!(3).powf(dd!(3.3));
let expected = dd!("37.540507598529552193101865954634");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-28));

pub fn recip(self) -> Double

Calculates the reciprocal of $x$, $\frac{1}{x}$, where $x$ is self.

Examples

let x = Double::PI.recip();
let expected = dd!("0.31830988618379067153776752674503");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

impl Double

pub fn rem_euclid(self, rhs: Double) -> Double

Calculates the least non-negative remainder of $x \div y$, where $x$ is self and $y$ is the argument. This is $x \mod y$ using the Euclidean definition of the modulo operation.

In principle, the returned value $r$ will satisfy $0 \le r < |y|$; it will be non-negative and less than the absolute value of the argument. However, in extreme cases, $r$ may equal $|y|$ because of floating-point rounding error.

Examples

let x = dd!(21);
let y = dd!(4);

assert!(x.rem_euclid(y) == dd!(1));
assert!((-x).rem_euclid(y) == dd!(3));
assert!(x.rem_euclid(-y) == dd!(1));
assert!((-x).rem_euclid(-y) == dd!(3));

pub fn div_euclid(self, rhs: Double) -> Double

Calculates the quotient in the Euclidean division $x \div y$, where $x$ is self and $y$ is the argument.

The Euclidean quotient is x / y rounded to the integer n such that x >= n * y. That makes it the counterpart to rem_euclid in the same way that regular division is the counterpart to the % operator; the answer n that it computes is the integer such that x = n * y + x.rem_euclid(y).

Examples

let x = dd!(21);
let y = dd!(4);

assert!(x.div_euclid(y) == dd!(5));
assert!((-x).div_euclid(y) == dd!(-6));
assert!(x.div_euclid(-y) == dd!(-5));
assert!((-x).div_euclid(-y) == dd!(6));

impl Double

pub fn sinh_cosh(self) -> (Double, Double)

Simultaneously computes the hyperbolic sine and cosine of $x$, $\sinh x$ and $\cosh x$, where $x$ is self.

The domain of this function is $(-\infin, \infin)$, and the range is $(-\infin, \infin)$ for the first component of the answer (the hyperbolic sine) and $[1, \infin)$ for the second (the hyperbolic cosine).

This method is more efficient to run than sinh and cosh individually and is useful when both numbers are needed.

Examples

let (sin_h, cos_h) = dd!(1).sinh_cosh();
let esin = dd!("1.1752011936438014568823818505956");
let ecos = dd!("1.5430806348152437784779056207571");

let diff1 = (sin_h - esin).abs();
let diff2 = (cos_h - ecos).abs();

assert!(diff1 < dd!(1e-30));
assert!(diff2 < dd!(1e-30));

pub fn sinh(self) -> Double

Computes the hyperbolic sine of $x$, $\sinh x$, where $x$ is self.

The domain and range of this function are both $(-\infin, \infin)$. Large values will start to cause a loss of precision; by the time the number is $\pm130$ or so, precision is down to 29 digits.

Examples

let x = dd!(1).sinh();
let expected = dd!("1.1752011936438014568823818505956");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn cosh(self) -> Double

Computes the hyperbolic cosine of $x$, $\cosh x$, where $x$ is self.

The domain of this function is $(-\infin, \infin)$, and the range is $[1, \infin)$.

Examples

let x = dd!(1).cosh();
let expected = dd!("1.5430806348152437784779056207571");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn tanh(self) -> Double

Computes the hyperbolic tangent of $x$, $\tanh x$, where $x$ is self.

The domain of this function is $(-\infin, \infin)$, and the range is $(-1, 1)$.

Examples

let x = dd!(1).tanh();
let expected = dd!("0.76159415595576488811945828260479");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn asinh(self) -> Double

Calculates the inverse hyperbolic sine of $x$, $\sinh^{-1} x$, where $x$ is self.

The domain and range of this function are both $(-\infin, \infin)$.

Examples

let x = dd!(1.5).asinh();
let expected = dd!("1.1947632172871093041119308285191");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn acosh(self) -> Double

Calculates the inverse hyperbolic cosine of $x$, $\cosh^{-1} x$, where $x$ is self.

The domain of the function is $[1, \infin)$ and the range is $[0, \infin)$. Any argument outside the range will result in NAN.

Examples

let x = dd!(1.5).acosh();
let expected = dd!("0.96242365011920689499551782684874");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn atanh(self) -> Double

Calculates the inverse hyperbolic tangent of $x$, $\tanh^{-1} x$, where $x$ is self.

The domain of the function is $(-1, 1)$ and the range is $(-\infin, \infin)$. Any argument whose absolute value is greater than or equal to 1 will result in NAN.

Examples

let x = dd!(0.5).atanh();
let expected = dd!("0.54930614433405484569762261846126");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

impl Double

pub fn abs(self) -> Double

Calculates the absolute value of $x$, $|x|$, where $x$ is self. The absolute value of $x$ is simply the same value as $x$, but with the opposite sign if $x$ is negative.

Examples

assert!(dd!(3).abs() == dd!(3));
assert!(dd!(-3).abs() == dd!(3));

pub fn floor(self) -> Double

Calculates the floor of $x$, $\lfloor{x}\rfloor$, where $x$ is self.

The floor of $x$ is the largest integer value less than or equal to $x$. This means that the floor of a negative number will have an absolute value greater than that of the number itself.

Examples

let f = dd!(3.99);
let g = dd!(3.0);
let h = dd!(-3.99);

assert!(f.floor() == dd!(3));
assert!(g.floor() == dd!(3));
assert!(h.floor() == dd!(-4));

pub fn ceil(self) -> Double

Calculates the ceiling of $x$, $\lceil{x}\rceil$, where $x$ is self.

The ceiling of $x$ is the smallest integer value greater than or equal to $x$. This means that the ceiling of a negative number will have an absolute value the same as (not greater than) that of the number itself.

Examples

let f = dd!(3.01);
let g = dd!(4.0);
let h = dd!(-3.01);

assert!(f.ceil() == dd!(4));
assert!(g.ceil() == dd!(4));
assert!(h.ceil() == dd!(-3));

pub fn round(self) -> Double

Calculates the rounded value of $x$, where $x$ is self.

The rounded value is the nearest integer to $x$. Halfway cases (i.e., numbers with a fractional portion of 0.5) are rounded away from 0, per the behavior of f64's round function.

Examples

let f = dd!(3.3);
let g = dd!(3.5);
let h = dd!(-3.3);

assert!(f.round() == dd!(3));
assert!(g.round() == dd!(4));
assert!(h.round() == dd!(-3));

pub fn trunc(self) -> Double

Returns the integer part of self. This integer part will be of the same sign as the original number.

Examples

let f = dd!(3.3);
let g = dd!(-3.7);

assert!(f.trunc() == dd!(3));
assert!(g.trunc() == dd!(-3));

pub fn fract(self) -> Double

Returns the fractional part of the self. This fractional part will be of the same sign as the original number.

Examples

let f = dd!(3.3);
let g = dd!(-3.7);

let fdiff = (f.fract() - dd!(0.3)).abs();
let gdiff = (g.fract() - dd!(-0.7)).abs();

assert!(fdiff < dd!(1e-30));
assert!(gdiff < dd!(1e-30));

pub fn signum(self) -> Double

Returns a number that represents the sign of self.

ONE if self is positive, including +0.0 and INFINITY

NEG_ONE if self is negative, including -0.0 and NEG_INFINITY

NAN if self is NAN

Examples

assert!(dd!(3.5).signum() == Double::ONE);
assert!(Double::NEG_INFINITY.signum() == Double::NEG_ONE);
assert!(Double::NAN.signum().is_nan());

pub fn classify(self) -> FpCategory

Returns self's floating point category.

The possible return values are the members of FpCategory, as follows:

FpCategory::Zero if the number is $\pm0$;

FpCategory::Infinite if the number is $\pm\infin$;

FpCategory::Nan if the number is not a number;

FpCategory::Subnormal if the number is $\pm$MIN_POSITIVE (numbers this small can be represented, but they lose some accuracy);

FpCategory::Normal if the number is anything else.

A Double can also register as FpCategory::Subnormal if it has a small enough negative exponent that the second component of the number is a subnormal number itself. This will typically happen around 1e-292 or so.

Examples

use std::num::FpCategory;

let num = dd!(12.4);
let inf = Double::INFINITY;

assert!(num.classify() == FpCategory::Normal);
assert!(inf.classify() == FpCategory::Infinite);

pub fn is_normal(self) -> bool

Returns true if self is neither zero, infinite, subnormal, or NaN.

Examples

let min = Double::MIN_POSITIVE;
let max = Double::MAX;
let lower = dd!(1e-308);
let zero = Double::ZERO;

assert!(min.is_normal());
assert!(max.is_normal());

assert!(!zero.is_normal());
assert!(!Double::NAN.is_normal());
assert!(!Double::INFINITY.is_normal());
// Values between `0` and `MIN_POSITIVE` are subnormal.
assert!(!lower.is_normal());

pub fn is_zero(self) -> bool

Returns true if self is either positive or negative zero.

Examples

assert!(Double::ZERO.is_zero());
assert!(Double::NEG_ZERO.is_zero());
assert!(!Double::PI.is_zero());

pub fn is_sign_negative(self) -> bool

Returns true if self is negative, including negative zero, negative infinity, and NaN with a negative sign bit.

Examples

assert!(Double::NEG_ZERO.is_sign_negative());
assert!(Double::NEG_INFINITY.is_sign_negative());
assert!(dd!(-7.0).is_sign_negative());
assert!(!Double::ZERO.is_sign_negative());
assert!(!dd!(7.0).is_sign_negative());

pub fn is_sign_positive(self) -> bool

Returns true if self is positive, including positive zero, positive infinity and NaN with a positive sign bit.

Examples

assert!(Double::ZERO.is_sign_positive());
assert!(Double::INFINITY.is_sign_positive());
assert!(dd!(7.0).is_sign_positive());
assert!(!Double::NEG_ZERO.is_sign_positive());
assert!(!dd!(-7.0).is_sign_positive());

pub fn is_nan(self) -> bool

Returns true if self is NaN.

This is the proper way to test for NaN because it cannot be done with an equality test (since NaN is not equal to itself).

Examples

assert!(Double::NAN.is_nan());
assert!(!dd!(7.0).is_nan());

pub fn is_infinite(self) -> bool

Returns true if self is positive or negative infinity.

Examples

assert!(Double::INFINITY.is_infinite());
assert!(Double::NEG_INFINITY.is_infinite());
assert!(!Double::NAN.is_infinite());
assert!(!dd!(7.0).is_infinite());

pub fn is_finite(self) -> bool

Returns true if self is neither infinite nor NaN.

Examples

assert!(!Double::INFINITY.is_finite());
assert!(!Double::NEG_INFINITY.is_finite());
assert!(!Double::NAN.is_finite());
assert!(dd!(7.0).is_finite());

pub fn is_subnormal(self) -> bool

Returns true if self has an absolute value of less than MIN_POSITIVE.

Numbers this small can be represented by floating point numbers, but they are not as accurate. This inaccuracy is inherent in the IEEE-754 format for 64-bit numbers; making a double-double out of an inaccurate number means the double-double is also going to be inaccurate.

Examples

assert!(!Double::PI.is_subnormal());
assert!(dd!(1e-308).is_subnormal());

impl Double

pub fn exp(self) -> Double

Computes the exponential function, $e^x$, where $x$ is self.

The result of this function grows rapidly. Once $x$ exceeds 708, the result is too large to represent with a Double; at that point the function begins to return INFINITY. The limit on the low end is less due to the fact that the second component needs to fit in an f64 rather than the first, along with extra bits used in argument reduction; this function begins to return 0 at -600.

As $x$ grows this function does lose a bit of precision. It's precise to at least 30 digits up to values of $-140 \le x \le 150$, and from then until the limits, it's precise to at least 29 digits.

Examples

let x = dd!(2).exp();
let expected = dd!("7.3890560989306502272304274605750057");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-29));

pub fn expm1(self) -> Double

Computes the exponential function of $x$ minus 1, $e^x - 1$, where $x$ is self.

While this function literally calculates the value returned by exp minus 1, it does this directly (rather than computing exp directly and then subtracting 1 from the answer). This is useful in the not-infrequent case where $x$ is very close to 0 and you have to subtract something near 1 from the answer (another example is x.exp() - x.cos() — for very small values of x, x.cos() is also very near 1). Since x.exp() is very close to 1 when x is very close to 0, this is subtracting something very close to 1 from something else very close to 1.

When this happens, subtracting 1 from exp suffers from catastrophic cancellation, a condition in which subtracting two numbers that are very close to each other can result in a huge loss of accuracy. Since expm1 does not perform this subtraction but instead computes ex - 1 directly, it does not suffer from this phenomenon.

This function will work over the same range as exp, but x.expm1() + 1 has no advantage over x.exp(). It's for that one very particular purpose.

Examples

// A value very close to 0.
let v_small = dd!(1e-28);
// The actual value of exp(v_small) - 1. The last digit is on the order of 1e-62.
let expected = dd!("1.0000000000000000000000000000500003e-28");

// First with exp, subtracting one from the result. The resulting precision loses at
// least 14 digits, making it no better than f64 precision.
let exp = v_small.exp() - Double::ONE;
let eps = (exp - expected).abs();
assert!(eps > dd!(1e-46));

// Next with expm1. Since this is calculated without subtracting one from anything,
// the catastrophic cancellation does not occur and full precision is maintained.
let expm1 = v_small.expm1();
let epsm1 = (expm1 - expected).abs();
assert!(epsm1 < dd!(1e-60));

pub fn ln(self) -> Double

Calculates the base-$e$ (natural) logarithm of $x$, $\ln x$ or $\log_e x$, where $x$ is self.

This calculation relies upon the exp calculation, in the opposite direction. A large positive logarithm, for example, will require the calculation of a large negative exponential.

For the same reasons that negative values of exp are limited to -600, the accurate results of this function are limited to the number whose logarithm is 600, which is around 2.65e261. Take care with this; unlike in exp, INFINITY is not returned. In that function, exceeding the maximum refers to actually overflowing an f64, which is appropriate to call INFINITY; here, it means 601.

Examples

let x = dd!(7).ln();
let expected = dd!("1.9459101490553133051053527434432");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-29));

pub fn ln1p(self) -> Double

Calculates the base-$e$ (natural) logarithm of 1 plus $x$, $\ln (1 + x)$ or $\log_e (1 + x)$, where $x$ is self.

This is the inverse of expm1 and arises from the same sorts of concerns. It isn't unusual to want to take logarithms of numbers very near 1, as the logarithm approaches 0 at that point. However, with finite-precision mathematics, the 1 itself severely limits the precision possible; the number 1.000000000000001 has 16 digits of precision, but most of them are taken up by placeholder zeros when we would prefer to have that precision available after the final 1.

ln1p allows that by letting the user pass in a number near 0 and having the algorithm add 1 to it internally, without causing the loss of precision. For example, the same 16-digit number above could be passed into ln1p as 0.000000000000001, a number with one digit of precision, leaving 15 more digits of precision availble after that final 1.

The algorithm for logarithms close to 1 is slower than that for the general logarithm, so this function delegates to ln if it can be done without losing precision. There is no advantage to using ln1p over ln except for computing logarithms of numbers very close to 1.

Examples

// A very small number
let v_small = dd!(1e-200);
// The actual value for the base-e logarithm of 1 + 1e-200
let expected = dd!("1.0000000000000000000000000000000004e-200");

// First with ln, computing the logarithm of 1 + 1e-200. This loses at least 14
// digits of precision because of the precision-damping effect of adding the one
// before the computation.
let ln = (Double::ONE + v_small).ln();
let lneps = (ln - expected).abs();
assert!(lneps > dd!(1e-217));

// Now with ln1p, computing the same logarithm. This does not suffer from the same
// effect from the 1, and the answer retains full precision.
let ln1p = v_small.ln1p();
let ln1peps = (ln1p - expected).abs();
assert!(ln1peps < dd!(1e-231));

pub fn log10(self) -> Double

Calculates the base-10 logarithm of $x$, $\log_{10} x$, where $x$ is self.

As with ln, this has an upper usable range less than the size of the numbers themselves. In this case, that upper limit is around 1e261. Over this number, the output is not reliable, but it does not return INFINITY because the number 261 is so plainly not infinite.

Examples

let x = Double::E.log10();
let expected = dd!("0.434294481903251827651128918916605");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn log2(self) -> Double

Calculates the base-2 logarithm of $x$, $\log_2 x$, where $x$ is self.

Since 2 is smaller than $e$, this function is constrained even more than ln. It will start returning NEG_INFINITY at around 1e-213 and will start to fail on the positive side at around 2.6e180.

Examples

let x = dd!(10).log2();
let expected = dd!("3.32192809488736234787031942948939");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-29));

pub fn log(self, b: Double) -> Double

Calculates the base-$b$ logarithm of $x$, $\log_b x$, where $x$ is self and $b$ is the argument.

This function will have limits at extreme arguments like the other logarithm functions. The difference is that those limits will depend on the base argument.

If the goal is to calculate the base-$e$, base-2, or base-10 logarithms of $x$, the specialized functions for those purposes (ln, log2, and log10 respectively) will be more efficient.

Examples

let x = dd!(10).log(dd!(7.0));
let expected = dd!("1.18329466245493832681792856164686");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-29));

impl Double

pub fn sin_cos(self) -> (Double, Double)

Simultaneously computes the sine and cosine of $x$, $\sin x$ and $\cos x$, where $x$ is self. This is more efficient than calling the separate sin and cos functions if you need both numbers.

Examples

let x = Double::PI / dd!(4);
let (sin_x, cos_x) = x.sin_cos();

let diff_sin = (sin_x - x.sin()).abs();
let diff_cos = (cos_x - x.cos()).abs();

assert!(diff_sin < dd!(1e-30));
assert!(diff_cos < dd!(1e-30));

pub fn sin(self) -> Double

Computes the sine of $x$, $\sin x$, where $x$ is self.

The domain of this function is $(-\infin, \infin)$, and the range is $[-1, 1]$.

Examples

let x = (Double::PI / dd!(2)).sin();
let expected = dd!(1);

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn cos(self) -> Double

Computes the cosine of $x$, $\cos x$, where $x$ is self.

The domain of this function is $(-\infin, \infin)$, and the range is $[-1, 1]$.

Examples

let x = (Double::PI / dd!(2)).cos();
let expected = dd!(0);

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn tan(self) -> Double

Computes the tangent of $x$, $\tan x$, where $x$ is self.

The domain and range of this function are both $(-\infin, \infin)$.

Examples

let x = (Double::PI / dd!(4)).tan();
let expected = dd!(1);

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn atan2(self, other: Double) -> Double

Computes the 2-argument inverse tangent of $x$ and $y$, $\arctan(y, x)$, where $x$ is the argument and $y$ is self.

The single-argument atan function always returns values in either the first (0 to $\pi$/2) or fourth (0 to $-\pi$/2) quadrants. However, first-quadrant results repeat themselves in the third quadrant, and fourth-quadrant results repeat themselves in the second. For example, $\tan \frac{\pi}{4}$ is 1, but so is $\tan \frac{3\pi}{4}$. Single-argument atan cannot distinguish between these two possibilities, so it always returns the one in the range $[-\frac{\pi}{2}, \frac{\pi}{2}]$.

The double-argument atan2 can return either, depending on the arguments. It essentially returns the angle between the positive x-axis and the point $(x, y)$, where $y$ is self and $x$ is the argument. Therefore dd!(1).atan2(dd!(1)) returns $\frac{\pi}{4}$ (first quadrant), but flipping both signs to dd!(-1).atan2(dd!(-1)) gives the $-\frac{3\pi}{4}$ result (third quadrant).

This function extends the range of the result to $[-\pi, \pi]$.

Because this function deals with angles around the origin and Cartesian coordinates, it's very useful for converting between Cartesian and polar coordinates.

Examples

let pi = Double::PI;

// -π/4 radians (45 degrees clockwise)
let x1 = dd!(3);
let y1 = dd!(-3);
let expected1 = -Double::FRAC_PI_4;

// 3π/4 radians (135 degrees counter-clockwise)
let x2 = dd!(-3);
let y2 = dd!(3);
let expected2 = Double::FRAC_3_PI_4;

let diff1 = (y1.atan2(x1) - expected1).abs();
let diff2 = (y2.atan2(x2) - expected2).abs();

assert!(diff1 < dd!(1e-30));
assert!(diff2 < dd!(1e-30));

pub fn asin(self) -> Double

Computes the inverse sine of $x$, $\sin^{-1} x$, where $x$ is self.

The domain of this function is $[-1, 1]$ while the range is $[-\frac{\pi}{2}, \frac{\pi}{2}]$. Arguments outside of this domain will result in NAN.

Examples

let x = dd!(1).asin();
let expected = Double::PI / dd!(2);  // π/2

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn acos(self) -> Double

Computes the inverse cosine of $x$, $\cos^{-1} x$, where $x$ is self.

The domain of this function is $[-1, 1]$ and the range is $[0, \pi]$. Arguments outside of the domain will result in NAN.

Examples

let x = dd!(1).acos();
let expected = dd!(0);

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn atan(self) -> Double

Computes the inverse tangent of $x$, $\tan^{-1} x$, where $x$ is self.

This is the single-argument version of this function (atan2 is the two-argument inverse tangent). Though inverse tangents have multiple answers, this function will always return the one with the lowest absolute value. x.atan() is the same as x.atan2(Double::ONE).

The domain of this function is $(-\infin, \infin)$ and the range is $[-\frac{\pi}{2}, \frac{\pi}{2}]$.

Examples

let x = dd!(1).atan();
let expected = Double::PI / dd!(4);  // π/4

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

impl Double

pub const fn new(a: f64, b: f64) -> Double

Creates a Double with the two arguments as the internal components.

Be sure you know what you're doing if you use this function. It does not normalize its components, meaning that if they aren't already normalized by the caller, this number will not work the way one would expect (it'll fail equality tests that it should pass, it may be classified incorrectly, etc.).

This function is primarily for creating constants where the normalization is obviously unnecessary. For example, if a Double version of the number 10 is needed, Double::new(10.0, 0.0) is a good way to do it in order to save the cost of the normalization that is obviously not needed.

Examples

let d = Double::new(0.0, 0.0);
assert!(d.is_zero());