Struct qd::Double

pub struct Double(_, _);

A 128-bit floating-point number implemented as the unevaluated sum of two 64-bit floating-point numbers. Discarding the bits used for exponents, this makes for about 106 bits of mantissa accuracy, or around 31 decimal digits.

There are several ways to create a new Double:

calling the new function

calling from with a primitive number (except for u128 and i128) or a string

calling parse on a string (or equivalently using from_str)

using the dd! macro

What kind of number you actually end up getting depends on the method called to get it. new will not normalize its result. This means that the arguments must be pre-normalized. from, parse, and dd! will both account for floating-point rounding error and produce normalized results.

The reason for these two different ways of going about creation is speed. If the number is already pre-computed to take normalization and error into account (as all of the constants in this library are), then new offers a way to avoid having to pay the efficiency cost of unnecessary normalization.

For the other methods, shortcuts can be taken if the input is a number and that number is dyadic (i.e., it can be represented in binary exactly, without rounding). In this case, from and dd! can also skip normalization and accounting for rounding, and they won't be much slower than new.

Parsing from strings or from numbers that are not dyadic cannot take these shortcuts. The results will be precise, but at the cost of speed.

See the module-level documentation for more information.

Implementations

impl Double

pub const RADIX: u32

The radix or base of the internal representation of Double. This is the same as the representation in the underlying f64.

pub const MANTISSA_DIGITS: u32

Number of significant digits in base 2.

pub const DIGITS: u32

Approximate number of significant digits in base 10.

pub const EPSILON: Double

Machine epsilon value for Double.

This is the difference between 1.0 and the next largest representable number.

pub const MIN: Double

Smallest finite Double value.

pub const MIN_POSITIVE: Double

Smallest positive normal Double value.

pub const MAX: Double

Largest finite Double value.

pub const MIN_EXP: i32

One greater than the minimum possible normal power of 2 exponent.

pub const MAX_EXP: i32

Maximum possible power of 2 exponent.

pub const MIN_10_EXP: i32

Minimum possible normal power of 10 exponent.

pub const MAX_10_EXP: i32

Maximum possible power of 10 exponent.

pub const NAN: Double

Not a Number (NaN).

pub const INFINITY: Double

Infinity (∞).

pub const NEG_INFINITY: Double

Negative infinity (-∞).

pub const ZERO: Double

Zero (0)

pub const NEG_ZERO: Double

Negative zero (-0)

pub const ONE: Double

One (1)

pub const NEG_ONE: Double

Negative one (-1)

pub const PI: Double

Archimedes' constant (π)

pub const TAU: Double

The full circle constant (τ), or 2π

pub const FRAC_PI_2: Double

π/2

pub const FRAC_PI_3: Double

π/3

pub const FRAC_PI_4: Double

π/4

pub const FRAC_PI_6: Double

π/6

pub const FRAC_PI_8: Double

π/8

pub const FRAC_PI_16: Double

π/16

pub const FRAC_3_PI_2: Double

3π/2

pub const FRAC_3_PI_4: Double

3π/4

pub const FRAC_5_PI_4: Double

5π/4

pub const FRAC_7_PI_4: Double

7π/4

pub const FRAC_1_PI: Double

1/π

pub const FRAC_2_PI: Double

2/π

pub const FRAC_2_SQRT_PI: Double

2/√π

pub const SQRT_2: Double

√2

pub const FRAC_1_SQRT_2: Double

1/√2

pub const E: Double

Euler's number (e)

pub const LOG2_10: Double

log₂ 10

pub const LOG2_E: Double

log₂ e

pub const LOG10_2: Double

log 2

pub const LOG10_E: Double

log e

pub const LN_2: Double

ln 2

pub const LN_10: Double

ln 10

impl Double

pub fn ldexp(self, n: i32) -> Double

Calculates x · 2n, where x is the Double and n is an integer.

Though this is not an everyday operation, it is often used in more advanced mathematical calculations (including several within this library). Therefore an implementation that is much more efficient than calculating it through multiplication and powi is offered despite it not being part of the f64 API.

Examples

let x = dd!(5);
assert!(x.ldexp(3) == dd!(40)); // 5 * 2^3

pub fn sqr(self) -> Double

Calculates the square of the Double.

This method takes advantage of optimizations in multiplication that are available when the two numbers being multiplied are the same, so it is more efficient than bare multiplication.

Examples

let x = dd!(3);
assert!(x.sqr() == x * x); // The left side is faster though

pub fn sqrt(self) -> Double

Calculates the square root of the Double.

Examples

let x = dd!(2).sqrt();
let diff = (x - Double::SQRT_2).abs();
assert!(diff < dd!(1e-30));

pub fn nroot(self, n: i32) -> Double

Calculates the nth root of the Double.

Examples

let x = dd!(2).nroot(4);
let expected = dd!("1.1892071150027210667174999705605");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn cbrt(self) -> Double

Calculates the cube root of the Double.

Examples

let x = dd!(2).cbrt();
let expected = dd!("1.2599210498948731647672106072782");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn powi(self, n: i32) -> Double

Calculates the Double raised to an integer power.

This function correctly handles the special inputs defined in IEEE 754. In particular:

x.powi(0) is 1 for any x (including 0, NaN, or infinity)

x.powi(n) is ±∞ for x = ±0 and any odd negative n

x.powi(n) is +∞ for x = ±0 and any even negative n

x.powi(n) is ±0 for x = ±0 and any odd positive n

x.powi(n) is +0 for x = ±0 and any even positive n

Examples

let x = dd!(3);
assert!(x.powi(3) == dd!(27));

pub fn powf(self, n: Double) -> Double

Calculates the Double raised to a Quad power.

In general, xn is equal to en ln x. This precludes raising a negative Double to a fractional or irrational power because ln x is undefined when x is negative. In that case, this function returns NAN.

It's actually more complex than that; if the exponent can be expressed as a fraction with an odd denominator, then there is an answer (a cube root, which is defined for negative numbers, is the same as a power of 1/3). Therefore, something like dd!(-4).powf(dd!(0.2)) should work, as 0.2 is a fraction with an odd denominator (1/5). However, it's impossible in general to tell whether a number is a fraction while using floating-point numbers, so no attempt is made to make this work. If you need a fifth root of -4, use dd!(-4).nroot(5).

Examples

let x = dd!(3).powf(dd!(3.3));
let expected = dd!("37.540507598529552193101865954634");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-28));

pub fn recip(self) -> Double

Calculates the reciprocal of the Double.

Examples

let x = Double::PI.recip();
let expected = dd!("0.31830988618379067153776752674503");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

impl Double

pub fn sinh_cosh(self) -> (Double, Double)

Simultaneously computes the hyperbolic sine and cosine (sinh and cosh) of the Double.

The domain of this function is (-∞, ∞), and the range is (-∞, ∞) for the first component of the answer (the hyperbolic sine) and [1, ∞) for the second (the hyperbolic cosine).

This method is more efficient to run than sinh and cosh individually and is useful when both numbers are needed.

Examples

let (sin_h, cos_h) = dd!(1).sinh_cosh();
let esin = dd!("1.1752011936438014568823818505956");
let ecos = dd!("1.5430806348152437784779056207571");

let diff1 = (sin_h - esin).abs();
let diff2 = (cos_h - ecos).abs();

assert!(diff1 < dd!(1e-30));
assert!(diff2 < dd!(1e-30));

pub fn sinh(self) -> Double

Computes the hyperbolic sine (sinh) of the Double.

The domain and range of this function are both (-∞, ∞). Large values will start to cause a loss of precision; by the time the number is ±130 or so, precision is down to 29 digits.

Examples

let x = dd!(1).sinh();
let expected = dd!("1.1752011936438014568823818505956");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn cosh(self) -> Double

Computes the hyperbolic cosine (cosh) of the Double.

The domain of this function is (-∞, ∞), and the range is [1, ∞).

Examples

let x = dd!(1).cosh();
let expected = dd!("1.5430806348152437784779056207571");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn tanh(self) -> Double

Computes the hyperbolic tangent (tanh) of the Double.

The domain of this function is (-∞, ∞), and the range is (-1, 1).

Examples

let x = dd!(1).tanh();
let expected = dd!("0.76159415595576488811945828260479");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn asinh(self) -> Double

Calculates the inverse hyperbolic sine (sinh-1) of the Double.

The domain and range of this function are both (-∞, ∞).

Examples

let x = dd!(1.5).asinh();
let expected = dd!("1.1947632172871093041119308285191");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn acosh(self) -> Double

Calculates the inverse hyperbolic cosine (cosh-1) of the Double.

The domain of the function is [1, ∞) and the range is [0, ∞). Any argument outside the range will result in NAN.

Examples

let x = dd!(1.5).acosh();
let expected = dd!("0.96242365011920689499551782684874");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn atanh(self) -> Double

Calculates the inverse hyperbolic tangent (tanh-1) of the Double.

The domain of the function is (-1, 1) and the range is (-∞, ∞). Any argument whose absolute value is greater than or equal to 1 will result in NAN.

Examples

let x = dd!(0.5).atanh();
let expected = dd!("0.54930614433405484569762261846126");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

impl Double

pub fn abs(self) -> Double

Calculates the absolute value of the Double.

Examples

assert!(dd!(3).abs() == dd!(3));
assert!(dd!(-3).abs() == dd!(3));

pub fn floor(self) -> Double

Returns the largest integer value less than or equal to the Double.

Examples

let f = dd!(3.99);
let g = dd!(3.0);
let h = dd!(-3.99);

assert!(f.floor() == dd!(3));
assert!(g.floor() == dd!(3));
assert!(h.floor() == dd!(-4));

pub fn ceil(self) -> Double

Returns the smallest integer value greater than or equal to the Double.

Examples

let f = dd!(3.01);
let g = dd!(4.0);
let h = dd!(-3.01);

assert!(f.ceil() == dd!(4));
assert!(g.ceil() == dd!(4));
assert!(h.ceil() == dd!(-3));

pub fn round(self) -> Double

Returns the nearest integer value to the Double. Half-way cases are rounded away from 0.0, per the behavior of f64's round method.

Examples

let f = dd!(3.3);
let g = dd!(3.5);
let h = dd!(-3.3);

assert!(f.round() == dd!(3));
assert!(g.round() == dd!(4));
assert!(h.round() == dd!(-3));

pub fn trunc(self) -> Double

Returns the integer part of the Double.

Examples

let f = dd!(3.3);
let g = dd!(-3.7);

assert!(f.trunc() == dd!(3));
assert!(g.trunc() == dd!(-3));

pub fn fract(self) -> Double

Returns the fractional part of the Double.

Examples

let f = dd!(3.3);
let g = dd!(-3.7);

let fdiff = (f.fract() - dd!(0.3)).abs();
let gdiff = (g.fract() - dd!(-0.7)).abs();

assert!(fdiff < dd!(1e-30));
assert!(gdiff < dd!(1e-30));

pub fn signum(self) -> Double

Returns a number that represents the sign of the Double.

1.0 if the number is positive, including +0.0 and INFINITY

-1.0 if the number is negative, including -0.0 and NEG_INFINITY

NAN if the number is NAN

Examples

assert!(dd!(3.5).signum() == Double::ONE);
assert!(Double::NEG_INFINITY.signum() == Double::NEG_ONE);
assert!(Double::NAN.signum().is_nan());

pub fn classify(self) -> FpCategory

Returns the floating point category of the Double.

The possible return values are the members of FpCategory, as follows:

FpCategory::Zero if the number is ±0;

FpCategory::Infinite if the number is ±∞;

FpCategory::Nan if the number is not a number;

FpCategory::Subnormal if the number is ±MIN_POSITIVE (numbers this small can be represented, but they lose some accuracy);

FpCategory::Normal if the number is anything else.

Examples

use std::num::FpCategory;

let num = dd!(12.4);
let inf = Double::INFINITY;

assert!(num.classify() == FpCategory::Normal);
assert!(inf.classify() == FpCategory::Infinite);

pub fn is_normal(self) -> bool

Returns true if the Double is neither zero, infinite, subnormal, or NaN.

Examples

let min = Double::MIN_POSITIVE;
let max = Double::MAX;
let lower = dd!(1e-308);
let zero = Double::ZERO;

assert!(min.is_normal());
assert!(max.is_normal());

assert!(!zero.is_normal());
assert!(!Double::NAN.is_normal());
assert!(!Double::INFINITY.is_normal());
// Values between `0` and `MIN_POSITIVE` are subnormal.
assert!(!lower.is_normal());

pub fn is_zero(self) -> bool

Returns true if the Double is either positive or negative zero.

Examples

assert!(Double::ZERO.is_zero());
assert!(Double::NEG_ZERO.is_zero());
assert!(!Double::PI.is_zero());

pub fn is_sign_negative(self) -> bool

Returns true if the Double is negative, including negative zero, negative infinity, and NaN with a negative sign bit.

Examples

assert!(Double::NEG_ZERO.is_sign_negative());
assert!(Double::NEG_INFINITY.is_sign_negative());
assert!(dd!(-7.0).is_sign_negative());
assert!(!Double::ZERO.is_sign_negative());
assert!(!dd!(7.0).is_sign_negative());

pub fn is_sign_positive(self) -> bool

Returns true if the Double is positive, including positive zero, positive infinity and NaN with a positive sign bit.

Examples

assert!(Double::ZERO.is_sign_positive());
assert!(Double::INFINITY.is_sign_positive());
assert!(dd!(7.0).is_sign_positive());
assert!(!Double::NEG_ZERO.is_sign_positive());
assert!(!dd!(-7.0).is_sign_positive());

pub fn is_nan(self) -> bool

Returns true if the Double is NaN.

Examples

assert!(Double::NAN.is_nan());
assert!(!dd!(7.0).is_nan());

pub fn is_infinite(self) -> bool

Returns true if the Double is positive or negative infinity.

Examples

assert!(Double::INFINITY.is_infinite());
assert!(Double::NEG_INFINITY.is_infinite());
assert!(!Double::NAN.is_infinite());
assert!(!dd!(7.0).is_infinite());

pub fn is_finite(self) -> bool

Returns true if the Double is neither infinite nor NaN.

Examples

assert!(!Double::INFINITY.is_finite());
assert!(!Double::NEG_INFINITY.is_finite());
assert!(!Double::NAN.is_finite());
assert!(dd!(7.0).is_finite());

pub fn is_subnormal(self) -> bool

Returns true if the Double has an absolute value of less than MIN_POSITIVE.

Numbers this small can be represented by floating point numbers, but they are not as accurate. This inaccuracy is inherent in the IEEE-754 format for 64-bit numbers; making a double-double out of an inaccurate number means the double-double is also going to be inaccurate.

Examples

assert!(!Double::PI.is_subnormal());
assert!(dd!(1e-308).is_subnormal());

impl Double

pub fn exp(self) -> Double

Computes the exponential function, ex, where x is this Double.

The result of this function grows rapidly. Once x exceeds 708, the result is too large to represent with a Double; at that point the function begins to return INFINITY. The limit on the low end is less due to the fact that the second component needs to fit in an f64 rather than the first, along with extra bits used in argument reduction; this function begins to return 0 at -600.

As x grows this function does lose a bit of precision. It's precise to at least 30 digits up to values of -140 <= x <= 150, and from then until the limits, it's precise to at least 29 digits.

Examples

let x = dd!(2).exp();
let expected = dd!("7.3890560989306502272304274605750057");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-29));

pub fn ln(self) -> Double

Calculates the natural logarithm, loge, of the Double.

This calculation relies upon the exp calculation, in the opposite direction. A large positive logarithm, for example, will require the calculation of a large negative exponential.

For the same reasons that negative values of exp are limited to -600, the accurate results of this function are limited to the number whose logarithm is 600, which is around 2.65 × 10261. Take care with this; unlike in exp, INFINITY is not returned. In that function, exceeding the maximum refers to actually overflowing an f64, which is appropriate to call INFINITY; here, it means 601.

Examples

let x = dd!(7).ln();
let expected = dd!("1.9459101490553133051053527434432");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-29));

pub fn log10(self) -> Double

Calculates the base-10 logarithm, log10, of the Double.

As with ln, this has an upper usable range less than the size of the numbers themselves. In this case, that upper limit is around 10261. Over this number, the output is not reliable, but it does not return INFINITY because the number 261 is so plainly not infinite.

Examples

let x = Double::E.log10();
let expected = dd!("0.434294481903251827651128918916605");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn log2(self) -> Double

Calculates the base-2 logarithm, log2, of the Double.

Since 2 is smaller than e, this function is constrained even more than ln. It will start returning NEG_INFINITY at around 10-213 and will start to fail on the positive side at around 2.6 × 10180.

Examples

let x = dd!(10).log2();
let expected = dd!("3.32192809488736234787031942948939");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-29));

pub fn log(self, b: Double) -> Double

Calculates the base b logarithm (logb) of the Double.

This function will have limits at extreme arguments like the other logarithm functions. The difference is that those limits will depend on the base argument.

If the goal is to calculate the base e, base 2, or base 10 logarithms of self, the specialized functions for those purposes(ln, log2, and log10 respectively) will be more efficient.

Examples

let x = dd!(10).log(dd!(7.0));
let expected = dd!("1.18329466245493832681792856164686");

let diff = (x - expected).abs();
assert!(diff < dd!(1e-29));

impl Double

pub fn sin_cos(self) -> (Double, Double)

Simultaneously computes the sine (sin) and the cosine (cos) of the Double. This is more efficient than calling the separate sin and cos functions if you need both numbers.

Examples

let x = Double::PI / dd!(4);
let (sin_x, cos_x) = x.sin_cos();

let diff_sin = (sin_x - x.sin()).abs();
let diff_cos = (cos_x - x.cos()).abs();

assert!(diff_sin < dd!(1e-30));
assert!(diff_cos < dd!(1e-30));

pub fn sin(self) -> Double

Computes the sine (sin) of the Double.

The domain of this function is (-∞, ∞), and the range is [-1, 1].

Examples

let x = (Double::PI / dd!(2)).sin();
let expected = dd!(1);

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn cos(self) -> Double

Computes the cosine (cos) of the Double.

The domain of this function is (-∞, ∞), and the range is [-1, 1].

Examples

let x = (Double::PI / dd!(2)).cos();
let expected = dd!(0);

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn tan(self) -> Double

Computes the tangent (tan) of the Double.

The domain and range of this function are both (-∞, ∞).

Examples

let x = (Double::PI / dd!(4)).tan();
let expected = dd!(1);

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn atan2(self, other: Double) -> Double

Computes the 2-argument inverse tangent (tan-1) of this Double and another Double.

The single-argument atan function always returns values in either the first (0 to π/2) or fourth (0 to -π/2) quadrants. However, first-quadrant results repeat themselves in the third quadrant, and fourth-quadrant results repeat themselves in the second. For example, the tangent of π/4 is 1, but so is the tangent of -3π/4. Single-argument atan cannot distinguish between these two possibilities, so it always returns the one in the range [-π/2, π/2].

The double-argument atan2 can return either, depending on the arguments. It essentially returns the angle between the positive x-axis and the point (x, y), where y is the Double that atan2 is called on and x is the argument. Therefore Double::ONE.atan2(Double::ONE) is π/4 (first quadrant), but flipping both signs to (Double::NEG_ONE).atan2(Double::NEG_ONE) gives the -3π/4 result (third quadrant).

This function extends the range of the result to [-π, π].

Because this function deals with angles around the origin and Cartesian coordinates, it's very useful for converting between Cartesian and polar coordinates.

Examples

let pi = Double::PI;

// -π/4 radians (45 degrees clockwise)
let x1 = dd!(3);
let y1 = dd!(-3);
let expected1 = -Double::FRAC_PI_4;

// 3π/4 radians (135 degrees counter-clockwise)
let x2 = dd!(-3);
let y2 = dd!(3);
let expected2 = Double::FRAC_3_PI_4;

let diff1 = (y1.atan2(x1) - expected1).abs();
let diff2 = (y2.atan2(x2) - expected2).abs();

assert!(diff1 < dd!(1e-30));
assert!(diff2 < dd!(1e-30));

pub fn asin(self) -> Double

Computes the inverse sine (sin-1) of the Double. The domain of this function is [-1, 1] while the range is [-π/2, π/2]. Arguments outside of this domain will result in NAN.

Examples

let x = dd!(1).asin();
let expected = Double::PI / dd!(2);  // π/2

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn acos(self) -> Double

Computes the inverse cosine (cos-1) of the Double. The domain of this function is [-1, 1] and the range is [0, π]. Arguments outside of the domain will result in NAN.

Examples

let x = dd!(1).acos();
let expected = dd!(0);

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

pub fn atan(self) -> Double

Computes the inverse tangent (tan-1) of the Double. The domain of this function is [-∞, ∞] and the range is [-π/2, π/2].

Examples

let x = dd!(1).atan();
let expected = Double::PI / dd!(4);  // π/4

let diff = (x - expected).abs();
assert!(diff < dd!(1e-30));

impl Double

pub const fn new(a: f64, b: f64) -> Double

Creates a Double with the two arguments as the internal components.

Be sure you know what you're doing if you use this function. It does not normalize its components, meaning that if they aren't already normalized by the caller, this number will not work the way one would expect (it'll fail equality tests that it should pass, it may be classified incorrectly, etc.).

This function is primarily for creating constants where the normalization is obviously unnecessary. For example, if a Double version of the number 10 is needed, Double::new(10.0, 0.0) is a good way to do it in order to save the cost of the normalization that is obviously not needed.

Examples

let d = Double::new(0.0, 0.0);
assert!(d.is_zero());