Struct qd::Quad

pub struct Quad(_, _, _, _);

A 256-bit floating-point number implemented as the unevaluated sum of four 64-bit floating-point numbers. Discarding the bits used for exponents, this makes for about 212 bits of mantissa accuracy, or around 63 decimal digits.

There are several ways to create a new Quad:

calling the new or raw functions

calling from with a primitive number or a string

calling parse on a string (or equivalently using from_str)

using the qd! macro

What kind of number you actually end up getting depends on the method called to get it.

raw will not normalize its result. This is for speed, but it means that the arguments must be pre-normalized.

new will normalize its result but will not account for floating-point rounding error. f64s passed to this functions are assumed to be exactly what's desired, including the rounding error.

from, parse, and qd! will both account for floating-point rounding error and produce normalized results. This is the slowest of the three choices but also the most accurate.

See the module-level documentation for more information.

Implementations

impl Quad

pub fn ldexp(self, n: i32) -> Quad

Calculates x · 2n, where x is the Quad and n is an integer.

Though this is not an everyday operation, it is often used in more advanced mathematical calculations (including several within this library). Therefore an implementation that is much more efficient than calculating it through multiplication and powi is offered despite it not being part of the f64 API.

Examples

let x = qd!(3);
assert!(x.ldexp(3) == qd!(24)); // 3 * 2^3

pub fn mul_pwr2(self, n: f64) -> Quad

Mutiplies the Quad by an f64 that must be a power of 2.

Like ldexp, this function is not present in the f64 API. However, it is much faster than regular multiplication for its very particular use. A regular multiplication between two Quads takes around 150 floating-point operations (depending on the numbers); this function takes 4.

The caveat is that this function will work only when n is a power of two (including negative powers, so 0.5, 0.25, etc. will also work). If this function is used with a number that is not a power of two, it will still return a result but that result will be wrong. Use this function only when you are 100% certain that the argument will be a power of two.

mul_pwr2 is primarily useful for implementing mathematical algorithms. It is used extensively in this library.

Examples

let x = Quad::PI.mul_pwr2(0.5); // MUCH faster than Quad::PI / 2
let expected = qd!("1.570796326794896619231321691639751442098584699687552910487472296");
 
let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn sqr(self) -> Quad

Calculates the square of the Quad.

This method takes advantage of optimizations in multiplication that are available when the two numbers being multiplied are the same, so it is more efficient than bare multiplication.

Examples

let x = qd!(3);
assert!(x.sqr() == x * x); // The left side is faster though

pub fn sqrt(self) -> Quad

Calculates the square root of the Quad.

Examples

let x = qd!(2).sqrt();
let expected = qd!("1.414213562373095048801688724209698078569671875376948073176679738");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn nroot(self, n: i32) -> Quad

Calculates the nth root of the Quad.

Examples

let x = qd!(2).nroot(4);
let expected = qd!("1.189207115002721066717499970560475915292972092463817413019002225");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn cbrt(self) -> Quad

Calculates the cube root of the Quad.

Examples

let x = qd!(2).cbrt();
let expected = qd!("1.259921049894873164767210607278228350570251464701507980081975112");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn powi(self, n: i32) -> Quad

Calculates the Quad raised to an integer power.

This function correctly handles the special inputs defined in IEEE 754. In particular:

x.powi(0) is 1 for any x (including 0, NaN, or infinity)

x.powi(n) is ±∞ for x = ±0 and any odd negative n

x.powi(n) is +∞ for x = ±0 and any even negative n

x.powi(n) is ±0 for x = ±0 and any odd positive n

x.powi(n) is +0 for x = ±0 and any even positive n

Examples

let x = qd!(3);
assert!(x.powi(3) == qd!(27));

pub fn powf(self, n: Quad) -> Quad

Calculates the Quad raised to a Quad power.

This function is implemented using the logarithm of the number being raised, which means it will not work for negatives even though raising a negative number to a non-integer power is defined. It has been extended to handle zero in accordance with the IEEE 754 specification.

It is possible that a new algorithm will eventually remove this restriction, though this is a surprisingly hard problem (see this libm implementation, for example).

Examples

let x = qd!(3).powf(qd!(3.3));
let expected = qd!("37.54050759852955219310186595463382927684873090166843452920390518");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn recip(self) -> Quad

Calculates the reciprocal of the Quad.

Examples

let x = Quad::PI.recip();
let expected = qd!("0.3183098861837906715377675267450287240689192914809128974953346881");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

impl Quad

pub const RADIX: u32

The radix or base of the internal representation of Double. This is the same as the representation in the underlying f64.

pub const MANTISSA_DIGITS: u32

Number of significant digits in base 2.

pub const DIGITS: u32

Approximate number of significant digits in base 10.

pub const EPSILON: Quad

Machine epsilon value for Quad.

This is the difference between 1.0 and the next largest representable number.

pub const MIN: Quad

Smallest finite Quad value.

pub const MIN_POSITIVE: Quad

Smallest positive normal Quad value.

pub const MAX: Quad

Largest finite Quad value.

pub const MIN_EXP: i32

One greater than the minimum possible normal power of 2 exponent.

pub const MAX_EXP: i32

Maximum possible power of 2 exponent.

pub const MIN_10_EXP: i32

Minimum possible normal power of 10 exponent.

pub const MAX_10_EXP: i32

Maximum possible power of 10 exponent.

pub const NAN: Quad

Not a Number (NaN).

pub const INFINITY: Quad

Infinity (∞).

pub const NEG_INFINITY: Quad

Negative infinity (-∞).

pub const ZERO: Quad

Zero (0)

pub const NEG_ZERO: Quad

Negative zero (-0)

pub const ONE: Quad

One (1)

pub const NEG_ONE: Quad

Negative one (-1)

pub const PI: Quad

Archimedes' constant (π)

pub const MUL_2_PI: Quad

2π

pub const FRAC_PI_2: Quad

π/2

pub const FRAC_PI_3: Quad

π/3

pub const FRAC_PI_4: Quad

π/4

pub const FRAC_PI_6: Quad

π/6

pub const FRAC_PI_8: Quad

π/8

pub const FRAC_PI_16: Quad

π/16

pub const FRAC_3_PI_4: Quad

3π/4

pub const FRAC_1_PI: Quad

1/π

pub const FRAC_2_PI: Quad

2/π

pub const FRAC_2_SQRT_PI: Quad

2/√π

pub const SQRT_2: Quad

√2

pub const FRAC_1_SQRT_2: Quad

1/√2

pub const E: Quad

Euler's number (e)

pub const LOG2_10: Quad

log₂ 10

pub const LOG2_E: Quad

log₂ e

pub const LOG10_2: Quad

log 2

pub const LOG10_E: Quad

log e

pub const LN_2: Quad

ln 2

pub const LN_10: Quad

ln 10

impl Quad

pub fn sinh_cosh(self) -> (Quad, Quad)

Simultaneously computes the hyperbolic sine and cosine (sinh and cosh) of the Quad.

The domain of this function is (-∞, ∞), and the range is (-∞, ∞) for the first component of the answer (the hyperbolic sine) and [1, ∞) for the second (the hyperbolic cosine).

This method is more efficient to run than sinh and cosh individually and is useful when both numbers are needed.

Examples

let (sin_h, cos_h) = qd!(1).sinh_cosh();
let esin = qd!("1.175201193643801456882381850595600815155717981334095870229565413");
let ecos = qd!("1.543080634815243778477905620757061682601529112365863704737402215");

let diff1 = (sin_h - esin).abs();
let diff2 = (cos_h - ecos).abs();

assert!(diff1 < qd!(1e-60));
assert!(diff2 < qd!(1e-60));

pub fn sinh(self) -> Quad

Computes the hyperbolic sine (sinh) of the Quad.

The domain and range of this function are both (-∞, ∞).

Examples

let x = qd!(1).sinh();
let expected = qd!("1.175201193643801456882381850595600815155717981334095870229565413");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn cosh(self) -> Quad

Computes the hyperbolic cosine (cosh) of the Quad.

The domain of this function is (-∞, ∞), and the range is [1, ∞).

Examples

let x = qd!(1).cosh();
let expected = qd!("1.543080634815243778477905620757061682601529112365863704737402215");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn tanh(self) -> Quad

Computes the hyperbolic tangent (tanh) of the Quad.

The domain of this function is (-∞, ∞), and the range is (-1, 1).

Examples

let x = qd!(1).tanh();
let expected = qd!("0.7615941559557648881194582826047935904127685972579365515968105001");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn asinh(self) -> Quad

Calculates the inverse hyperbolic sine (sinh-1) of the Quad.

The domain and range of this function are both (-∞, ∞).

Examples

let x = qd!(1.5).asinh();
let expected = qd!("1.194763217287109304111930828519090523536162075153005429270680299");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn acosh(self) -> Quad

Calculates the inverse hyperbolic cosine (cosh-1) of the Quad.

The domain of the function is [1, ∞) and the range is [0, ∞). Any argument outside the range will result in NAN.

Examples

let x = qd!(1.5).acosh();
let expected = qd!("0.9624236501192068949955178268487368462703686687713210393220363377");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-30));

pub fn atanh(self) -> Quad

Calculates the inverse hyperbolic tangent (tanh-1) of the Quad.

The domain of the function is (-1, 1) and the range is (-∞, ∞). Any argument whose absolute value is greater than or equal to 1 will result in NAN.

Examples

let x = qd!(0.5).atanh();
let expected = qd!("0.5493061443340548456976226184612628523237452789113747258673471668");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

impl Quad

pub fn abs(self) -> Quad

Calculates the absolute value of the Quad.

Examples

assert!(qd!(3).abs() == qd!(3));
assert!(qd!(-3).abs() == qd!(3));

pub fn floor(self) -> Quad

Returns the largest integer value less than or equal to the Quad.

Examples

let f = qd!(3.99);
let g = qd!(3.0);

assert!(f.floor() == qd!(3));
assert!(g.floor() == qd!(3));

pub fn ceil(self) -> Quad

Returns the smallest integer value greater than or equal to the Quad.

Examples

let f = qd!(3.01);
let g = qd!(4.0);

assert!(f.ceil() == qd!(4));
assert!(g.ceil() == qd!(4));

pub fn round(self) -> Quad

Returns the nearest integer value to the Double. Half-way cases are rounded away from 0.0, per the behavior of f64's round method.

Examples

let f = qd!(3.3);
let g = qd!(-3.3);

assert!(f.round() == qd!(3));
assert!(g.round() == qd!(-3));

pub fn trunc(self) -> Quad

Returns the integer part of the Quad.

Examples

let f = qd!(3.3);
let g = qd!(-3.7);

assert!(f.trunc() == qd!(3));
assert!(g.trunc() == qd!(-3));

pub fn fract(self) -> Quad

Returns the fractional part of the Quad.

Examples

let f = qd!(3.3);
let g = qd!(-3.7);

let fdiff = (f.fract() - qd!(0.3)).abs();
let gdiff = (g.fract() - qd!(-0.7)).abs();

assert!(fdiff < qd!(1e-60));
assert!(gdiff < qd!(1e-60));

pub fn signum(self) -> Quad

Returns a number that represents the sign of the Quad.

1.0 if the number is positive, including +0.0 and INFINITY

-1.0 if the number is negative, including -0.0 and NEG_INFINITY

NAN if the number is NAN

Examples

assert!(qd!(3.5).signum() == Quad::ONE);
assert!(Quad::NEG_INFINITY.signum() == Quad::NEG_ONE);
assert!(Quad::NAN.signum().is_nan());

pub fn classify(self) -> FpCategory

Returns the floating point category of the Quad.

The possible return values are the members of FpCategory, as follows:

FpCategory::Zero if the number is ±0;

FpCategory::Infinite if the number is ±∞;

FpCategory::Nan if the number is not a number;

FpCategory::Subnormal if the number is ±MIN_POSITIVE (numbers this small can be represented, but they lose some accuracy);

FpCategory::Normal if the number is anything else.

Examples

use std::num::FpCategory;

let num = qd!(12.4);
let inf = Quad::INFINITY;

assert!(num.classify() == FpCategory::Normal);
assert!(inf.classify() == FpCategory::Infinite);

pub fn is_normal(self) -> bool

Returns true if the Quad is neither zero, infinite, subnormal, or NaN.

Examples

let min = Quad::MIN_POSITIVE;
let max = Quad::MAX;
let lower = qd!(1e-308);
let zero = Quad::ZERO;

assert!(min.is_normal());
assert!(max.is_normal());

assert!(!zero.is_normal());
assert!(!Quad::NAN.is_normal());
assert!(!Quad::INFINITY.is_normal());
// Values between `0` and `MIN_POSITIVE` are subnormal.
assert!(!lower.is_normal());

pub fn is_zero(self) -> bool

Returns true if the Quad is either positive or negative zero.

Examples

assert!(Quad::ZERO.is_zero());
assert!(Quad::NEG_ZERO.is_zero());
assert!(!Quad::PI.is_zero());

pub fn is_sign_negative(self) -> bool

Returns true if the Quad is negative, including negative zero, negative infinity, and NaN with a negative sign bit.

Examples

assert!(Quad::NEG_ZERO.is_sign_negative());
assert!(Quad::NEG_INFINITY.is_sign_negative());
assert!(qd!(-7.0).is_sign_negative());
assert!(!Quad::ZERO.is_sign_negative());
assert!(!qd!(7.0).is_sign_negative());

pub fn is_sign_positive(self) -> bool

Returns true if the Quad is positive, including positive zero, positive infinity and NaN with a positive sign bit.

Examples

assert!(Quad::ZERO.is_sign_positive());
assert!(Quad::INFINITY.is_sign_positive());
assert!(qd!(7.0).is_sign_positive());
assert!(!Quad::NEG_ZERO.is_sign_positive());
assert!(!qd!(-7.0).is_sign_positive());

pub fn is_nan(self) -> bool

Returns true if the Quad is NaN.

Examples

assert!(Quad::NAN.is_nan());
assert!(!qd!(7.0).is_nan());

pub fn is_infinite(self) -> bool

Returns true if the Quad is positive or negative infinity.

Examples

assert!(Quad::INFINITY.is_infinite());
assert!(Quad::NEG_INFINITY.is_infinite());
assert!(!Quad::NAN.is_infinite());
assert!(!qd!(7.0).is_infinite());

pub fn is_finite(self) -> bool

Returns true if the Quad is neither infinite nor NaN..

Examples

assert!(!Quad::INFINITY.is_finite());
assert!(!Quad::NEG_INFINITY.is_finite());
assert!(!Quad::NAN.is_finite());
assert!(qd!(7.0).is_finite());

pub fn is_subnormal(self) -> bool

Returns true if the Quad has an absolute value of less than MIN_POSITIVE.

Numbers this small can be represented by floating point numbers, but they are not as accurate. This inaccuracy is inherent in the IEEE-754 format for 64-bit numbers; making a double-double out of an inaccurate number means the double-double is also going to be inaccurate.

Examples

assert!(!Quad::PI.is_subnormal());
assert!(qd!(1e-308).is_subnormal());

impl Quad

pub fn exp(self) -> Quad

Computes the exponential function, ex, for the Quad.

Examples

let x = qd!(2.3).exp();
let expected = qd!("9.974182454814720739957615156908858001478701193684029563691421917");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn ln(self) -> Quad

Calculates the natural logarithm, loge, of the Quad.

Examples

let x = qd!(7).ln();
let expected = qd!("1.945910149055313305105352743443179729637084729581861188459390150");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn log10(self) -> Quad

Calculates log10 of the Quad.

Examples

let x = Quad::E.log10();
let expected = qd!("0.4342944819032518276511289189166050822943970058036665661144537832");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn log2(self) -> Quad

Calculates log2 of the Quad.

Examples

let x = qd!(10).log2();
let expected = qd!("3.321928094887362347870319429489390175864831393024580612054756396");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn log(self, b: f64) -> Quad

Calculates the base b logarithm (logb) of the Quad.

If the goal is to calculate the base e, base 2, or base 10 logarithms of self, the specialized functions for those purposes(ln, log2, and log10 respectively) will be more efficient.

Examples

let x = qd!(10).log(7.3);
let expected = qd!("1.158315209978887965104764376269736420106652944692834002126233653");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

impl Quad

pub fn sin_cos(self) -> (Quad, Quad)

Simultaneously computes the sine (sin) and the cosine (cos) of the Quad. This is more efficient than calling the separate sin and cos functions if you need both numbers.

Examples

let x = Quad::PI / qd!(4);
let (sin_x, cos_x) = x.sin_cos();

let diff_sin = (sin_x - x.sin()).abs();
let diff_cos = (cos_x - x.cos()).abs();

assert!(diff_sin < qd!(1e-60));
assert!(diff_cos < qd!(1e-60));

pub fn sin(self) -> Quad

Computes the sine (sin) of the Quad.

The domain of this function is (-∞, ∞), and the range is [-1, 1].

Examples

let x = qd!(1).sin();
let expected = qd!("0.8414709848078965066525023216302989996225630607983710656727517099884");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn cos(self) -> Quad

Computes the cosine (cos) of the Quad.

The domain of this function is (-∞, ∞), and the range is [-1, 1].

Examples

let x = qd!(1).cos();
let expected = qd!("0.5403023058681397174009366074429766037323104206179222276700972553787");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn tan(self) -> Quad

Computes the tangent (tan) of the Quad.

The domain and range of this function are both (-∞, ∞).

Examples

let x = qd!(1).tan();
let expected = qd!("1.557407724654902230506974807458360173087250772381520038383946606");

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn atan2(self, other: Quad) -> Quad

Computes the 2-argument inverse tangent (tan-1) of this Quad and another Quad.

The single-argument atan function always returns values in either the first (0 to π/2) or fourth (0 to -π/2) quadrants. However, first-quadrant results repeat themselves in the third quadrant, and fourth-quadrant results repeat themselves in the second. For example, the tangent of π/4 is 1, but so is the tangent of -3π/4. Single-argument atan cannot distinguish between these two possibilities, so it always returns the one in the range [-π/2, π/2].

The double-argument atan2 can return either, depending on the arguments. It essentially returns the angle between the positive x-axis and the point (x, y), where y is the Quad that atan2 is called on and x is the argument. Therefore Quad::ONE.atan2(Quad::ONE) is π/4 (first quadrant), but flipping both signs to (Quad::NEG_ONE).atan2(Quad::NEG_ONE) gives the -3π/4 result (third quadrant).

This function extends the range of the result to [-π, π].

Because this function deals with angles around the origin and Cartesian coordinates, it's very useful for converting between Cartesian and polar coordinates.

Examples

let pi = Quad::PI;

// -π/4 radians (45 degrees clockwise)
let x1 = qd!(3);
let y1 = qd!(-3);
let expected1 = -pi / qd!(4);

// 3π/4 radians (135 degrees counter-clockwise)
let x2 = qd!(-3);
let y2 = qd!(3);
let expected2 = Quad::from(0.75) * pi;

let diff1 = (y1.atan2(x1) - expected1).abs();
let diff2 = (y2.atan2(x2) - expected2).abs();

assert!(diff1 < qd!(1e-60));
assert!(diff2 < qd!(1e-60));

pub fn asin(self) -> Quad

Computes the inverse sine (sin-1) of the Quad. The domain of this function is [-1, 1] while the range is [-π/2, π/2]. Arguments outside of this domain will result in NAN.

Examples

let x = qd!(1).asin();
let expected = Quad::PI / qd!(2);  // π/2

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn acos(self) -> Quad

Computes the inverse cosine (cos-1) of the Quad. The domain of this function is [-1, 1] and the range is [0, π]. Arguments outside of the domain will result in NAN.

Examples

let x = qd!(1).acos();
let expected = qd!(0);

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

pub fn atan(self) -> Quad

Computes the inverse tangent (tan-1) of the Quad. The domain of this function is [-∞, ∞] and the range is [-π/2, π/2].

Examples

let x = qd!(1).atan();
let expected = Quad::PI / qd!(4);  // π/4

let diff = (x - expected).abs();
assert!(diff < qd!(1e-60));

impl Quad

pub fn raw(a: f64, b: f64, c: f64, d: f64) -> Quad

Creates a Quad with the four arguments as the internal components.

Be sure you know what you're doing if you use this function. It does not normalize its components, meaning that if they aren't already normalized by the caller, this number will not work the way one would expect (it'll fail equality tests that it should pass, it may be classified incorrectly, etc.).

This function is primarily for creating constants where the normalization is obviously unnecessary. For example, if a Quad version of the number 10 is needed, Quad::raw(10.0, 0.0, 0.0, 0.0) is a good way to do it in order to save the cost of the normalization that is obviously not needed.

Examples

let q = Quad::raw(0.0, 0.0, 0.0, 0.0);
assert!(q.is_zero());

pub fn new(a: f64, b: f64, c: f64, d: f64) -> Quad

Creates a Quad by normalizing its four arguments.

This function normalizes the input arguments (if this is obviously unnecessary, use raw instead) and assigns the normalized values to the new Quad's components.

It's assumed that the four numbers passed in are exactly what's desired, and aside from normalization, they will not be manipulated further. That means that any floating-point rounding error will be retained. For instance, Quad::new(1.1, 0.0, 0.0, 0.0) actually produces the number 1.100000000000000088817841970012523233890533447265625. To account for that rounding error, use from or the qd! macro; qd!(1.1) is effectively the same as Quad::new(1.1, -8.881784197001253e-17, 4.930380657631324e-33, -2.7369110631344085e-49).

Examples

let d = Quad::new(2.0, 1.0, 3.0, -4.5);
assert!(d == qd!(1.5));